% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{as.data.frame.bertopic_r}
\alias{as.data.frame.bertopic_r}
\title{Coerce to data.frame}
\usage{
\method{as.data.frame}{bertopic_r}(x, ...)
}
\arguments{
\item{x}{A "bertopic_r" model.}

\item{...}{Unused.}
}
\value{
A data.frame equal to \code{\link[=bertopic_topics]{bertopic_topics()}}.
}
\description{
Coerce to data.frame
}
