% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-documents.R
\name{bert_document_info}
\alias{bert_document_info}
\title{BERTopic Documents}
\usage{
bert_document_info(
  obj,
  docs,
  document_name = NULL,
  exclude_list_columns = T,
  assign_to_environment = TRUE
)
}
\arguments{
\item{obj}{BERTopic model object.}

\item{docs}{Character vector of documents.}

\item{document_name}{Character. If not `NULL`, name for the text document column.}

\item{exclude_list_columns}{Logical. If `TRUE`, excludes list columns from output. Default `TRUE`.}

\item{assign_to_environment}{Logical. If TRUE, assigns result tibble to global environment. Default FALSE.}
}
\value{
A tibble with document information including topic assignments and representative documents.
}
\description{
BERTopic Documents
}
