% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_fit}
\alias{bert_fit}
\title{BERTopic Fit}
\usage{
bert_fit(obj, documents, embeddings = NULL, y = NULL, images = NULL)
}
\arguments{
\item{obj}{A BERTopic model object created by `bert_topic()`.}

\item{documents}{Character vector. A single document or list of documents to fit on.}

\item{embeddings}{Numeric matrix or `NULL`. Pre-trained document embeddings. These can be used instead of the sentence-transformer model. Default `NULL`.}

\item{y}{Integer vector or `NULL`. The target class for (semi)-supervised modeling. Use `-1` if no class for a specific instance is specified. Default `NULL`.}

\item{images}{List or `NULL`. A list of paths to the images to fit on or the images themselves. Default `NULL`.}
}
\value{
The fitted BERTopic model object.
}
\description{
Fit the models (BERT, UMAP, and HDBSCAN) on a collection of documents and generate topics.
}
