% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_fit_transform}
\alias{bert_fit_transform}
\title{BERTopic Fit and Transform}
\usage{
bert_fit_transform(obj, documents, embeddings = NULL, y = NULL)
}
\arguments{
\item{obj}{A BERTopic model object created by `bert_topic()`.}

\item{documents}{Character vector. A single document or list of documents to fit on.}

\item{embeddings}{Numeric matrix or `NULL`. Pre-trained document embeddings. These can be used instead of the sentence-transformer model. Default `NULL`.}

\item{y}{Integer vector or `NULL`. The target class for (semi)-supervised modeling. Use `-1` if no class for a specific instance is specified. Default `NULL`.}
}
\value{
An integer vector of assigned topic IDs for each document.
}
\description{
Fit the models on a collection of documents, generate topics, and return the docs with topics.
}
