% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_merge_models}
\alias{bert_merge_models}
\title{Merge multiple pre-trained BERTopic models into a single model}
\usage{
bert_merge_models(
  obj,
  models = NULL,
  min_similarity = 0.7,
  embedding_model = NULL
)
}
\arguments{
\item{obj}{A BERTopic model object created by `bert_topic()`.}

\item{models}{List of fitted BERTopic model objects to merge.}

\item{min_similarity}{Numeric. The minimum similarity score for when topics are merged. Default `0.7`.}

\item{embedding_model}{Python embedding model or `NULL`. Additionally load in an embedding model if necessary. Default `NULL`.}
}
\value{
The merged BERTopic model object with consolidated topics from all input models.
}
\description{
Merge multiple pre-trained BERTopic models into a single model
}
