% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-topic-modify.R
\name{bert_merge_topics}
\alias{bert_merge_topics}
\title{Merge Topics}
\usage{
bert_merge_topics(obj, docs, topics_to_merge = NULL)
}
\arguments{
\item{obj}{BERTopic model object.}

\item{docs}{Character vector of documents used in model fitting.}

\item{topics_to_merge}{List. Either a list of topic IDs or a list of lists of topic IDs to merge. For example: `list(1,2,3)` merges topics 1, 2, and 3; `list(list(1,2), list(3,4))` merges 1 and 2 separately from 3 and 4.}
}
\value{
The updated BERTopic model object with merged topics.
}
\description{
Merge Topics
}
