% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_parameters}
\alias{bert_parameters}
\title{BERTopic Parameters}
\usage{
bert_parameters(
  obj,
  deep = FALSE,
  return_tibble = FALSE,
  use_dfc_method = FALSE,
  return_attributes = FALSE
)
}
\arguments{
\item{obj}{A BERTopic model object or Python object with `get_params()` method.}

\item{deep}{Logical. If `TRUE` returns deep information. Default `FALSE`.}

\item{return_tibble}{Logical. If `TRUE` returns a `tibble`. Default `FALSE`.}

\item{use_dfc_method}{Logical. If `TRUE` uses `map_dfc` method for conversion. Default `FALSE`.}

\item{return_attributes}{Logical. If `TRUE` returns concatenated character vector of the parameters. Default `FALSE`.}
}
\value{
A list (default), tibble (if `return_tibble = TRUE`), or character vector (if `return_attributes = TRUE`) containing model parameters.
}
\description{
BERTopic Parameters
}
