% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_stop_words}
\alias{bert_stop_words}
\title{Bertopic Stopwords}
\usage{
bert_stop_words(extra_stop_words = NULL)
}
\arguments{
\item{extra_stop_words}{Character vector of additional stopwords to append.}
}
\value{
A character vector of stopwords including base words and any extra stopwords provided.
}
\description{
Bertopic Stopwords
}
