% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-io.R
\name{bert_stopwords}
\alias{bert_stopwords}
\title{Stopword List Generator}
\usage{
bert_stopwords(
  language = "english",
  is_lower_case = T,
  stopword_package_sources = NULL,
  extra_stop_words = NULL
)
}
\arguments{
\item{language}{Character. Language code. Default `"english"`.}

\item{is_lower_case}{Logical. If `TRUE`, converts stopwords to lowercase. Default `TRUE`.}

\item{stopword_package_sources}{Character vector. External stopword sources to combine. Default `NULL`.}

\item{extra_stop_words}{Character vector. Additional stopwords to include. Default `NULL`.}
}
\value{
A character vector of unique stopwords.
}
\description{
Stopword List Generator
}
