% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_topic}
\alias{bert_topic}
\title{Initiate BERT Topic Model}
\usage{
bert_topic(
  obj = NULL,
  umap_model = NULL,
  hdbscan_model = NULL,
  vectorizer_model = NULL,
  embedding_model = NULL,
  ctfidf_model = NULL,
  nr_topics = NULL,
  language = "english",
  representation_model = NULL,
  top_n_words = 10L,
  use_key_phrase_vectorizer = F,
  use_sklearn_vectorizer = F,
  is_lower_case = T,
  n_gram_range = list(1L, 1L),
  zeroshot_topic_list = NULL,
  zeroshot_min_similarity = 0.7,
  keyphrase_ngram_range = list(1L, 1L),
  min_topic_size = 10L,
  low_memory = F,
  exclude_stop_words = T,
  stopword_package_sources = NULL,
  extra_stop_words = NULL,
  calculate_probabilities = T,
  min_df = 1L,
  max_df = 1L,
  pos_pattern = "<J.*>*<N.*>+",
  seed_topic_list = NULL,
  verbose = T,
  vocabulary = NULL,
  workers = 6L,
  decay = NULL,
  delete_min_df = NULL,
  numba_threads = 1L,
  use_token_parallel = TRUE,
  set_attributes = TRUE
)
}
\arguments{
\item{obj}{Python BERTopic module object or `NULL` to import.}

\item{umap_model}{Python UMAP model or `NULL`. Pass in a UMAP model to be used instead of the default. NOTE: You can also pass in any dimensionality reduction algorithm as long as it has `.fit` and `.transform` functions. Default `NULL`.}

\item{hdbscan_model}{Python HDBSCAN model or `NULL`. Pass in an `hdbscan.HDBSCAN` model to be used instead of the default. NOTE: You can also pass in any clustering algorithm as long as it has `.fit` and `.predict` functions along with the `.labels_` variable. Default `NULL`.}

\item{vectorizer_model}{Python vectorizer model or `NULL`. Pass in a custom CountVectorizer instead of the default model. Default `NULL`.}

\item{embedding_model}{Python embedding model or `NULL`. Type of embedding model - either a model object or `NULL`. Options include \href{https://www.sbert.net/docs/pretrained_models.html}{sbert} and defaults to `"all-MiniLM-L6-v2"`. Default `NULL`.}

\item{ctfidf_model}{Python C-TF-IDF model or `NULL`. Pass in a custom ClassTfidfTransformer instead of the default model. Default `NULL`.}

\item{nr_topics}{Integer or `NULL`. Specifying the number of topics will reduce the initial number of topics to the value specified. This reduction can take a while as each reduction in topics (-1) activates a c-TF-IDF calculation. If `NULL`, no reduction is applied. Use `"auto"` to automatically reduce topics using HDBSCAN. Default `NULL`.}

\item{language}{Character. The main language used in your documents. The default sentence-transformers model for "english" is all-MiniLM-L6-v2. For a full overview of supported languages see `bertopic.backend.languages`. Select "multilingual" to load the paraphrase-multilingual-MiniLM-L12-v2 sentence-transformers model that supports 50+ languages. Default `"english"`.}

\item{representation_model}{Python representation model or `NULL`. The base representation model for fine-tuning topic representations. Default `NULL`.}

\item{top_n_words}{Integer. The number of words per topic to extract. Setting this too high can negatively impact topic embeddings as topics are typically best represented by at most 10 words. Default `10L`.}

\item{use_key_phrase_vectorizer}{Logical. If `TRUE` uses a keyphrase vectorizer. Default `FALSE`.}

\item{use_sklearn_vectorizer}{Logical. If `TRUE` uses scikit-learn vectorizer. Default `FALSE`.}

\item{is_lower_case}{Logical. If `TRUE` converts text to lowercase. Default `TRUE`.}

\item{n_gram_range}{List of integers. The n-gram range for the CountVectorizer. Advised to keep high values between 1 and 3. More would likely lead to memory issues. NOTE: This param will not be used if you pass in your own CountVectorizer. Default `list(1L, 1L)`.}

\item{zeroshot_topic_list}{List or `NULL`. List of candidate topic labels for zero-shot classification. Default `NULL`.}

\item{zeroshot_min_similarity}{Numeric. Minimum similarity score for zero-shot topic matching. Default `0.7`.}

\item{keyphrase_ngram_range}{List or `NULL`. Range for keyphrase n-grams. Default `list(1L, 1L)`.}

\item{min_topic_size}{Integer. The minimum size of the topic. Increasing this value will lead to a lower number of clusters/topics. NOTE: This param will not be used if you are not using HDBSCAN. Default `10L`.}

\item{low_memory}{Logical. Sets UMAP low memory to `TRUE` to make sure less memory is used. NOTE: This is only used in UMAP. For example, if you use PCA instead of UMAP this parameter will not be used. Default `FALSE`.}

\item{exclude_stop_words}{Logical. If `TRUE` excludes base stop words. Default `TRUE`.}

\item{stopword_package_sources}{Character vector or `NULL`. Options if not `NULL`: `c("snowball", "stopwords-iso", "smart", "nltk")`. Default `NULL`.}

\item{extra_stop_words}{Character vector or `NULL`. Additional stopwords to exclude. Default `NULL`.}

\item{calculate_probabilities}{Logical. If `TRUE` calculates the probabilities of all topics per document instead of the probability of the assigned topic per document. This could slow down the extraction of topics if you have many documents (> 100,000). Set this only to `TRUE` if you have a low amount of documents or if you do not mind more computation time. NOTE: If `FALSE` you cannot use the corresponding visualization method `visualize_probabilities`. Default `TRUE`.}

\item{min_df}{Integer. During fitting ignore keyphrases that have a document frequency strictly lower than the given threshold. This value is also called cut-off in the literature. Default `1L`.}

\item{max_df}{Integer. During fitting ignore keyphrases that have a document frequency strictly higher than the given threshold. Default `1L`.}

\item{pos_pattern}{Character. Position pattern for keyphrase extraction. Defaults to `"<J.*>*<N.*>+"`. See \href{https://maartengr.github.io/KeyBERT/guides/countvectorizer.html#languages}{Position Pattern via KeyBERT}, \href{https://keyphrase-vectorizers.readthedocs.io/en/latest/api.html}{KeyPhrase POS API}, and \href{https://github.com/TimSchopf/KeyphraseVectorizers#custom-pos-tagger}{Custom Tagger via KeyPhrase} for details.}

\item{seed_topic_list}{List or `NULL`. A list of seed words per topic to converge around. Default `NULL`.}

\item{verbose}{Logical. If `TRUE` verbose output is printed. Default `TRUE`.}

\item{vocabulary}{List or `NULL`. Custom vocabulary for the vectorizer. Default `NULL`.}

\item{workers}{Integer. Number of parallel workers for processing. Default `6L`.}

\item{decay}{Numeric or `NULL`. Decay parameter for online learning. Default `NULL`.}

\item{delete_min_df}{Logical or `NULL`. If `TRUE` deletes documents with frequency below `min_df`. Default `NULL`.}

\item{numba_threads}{Integer. Number of threads for Numba operations. Default `1L`.}

\item{use_token_parallel}{Logical. If `TRUE` enables parallel tokenization. Default `TRUE`.}

\item{set_attributes}{Logical. If `TRUE` sets attributes on the model object. Default `TRUE`.}
}
\value{
A Python BERTopic model object configured with the specified parameters.
}
\description{
Functions for unsupervised clustering algorithms.
}
\details{
\itemize{
\item \href{https://maartengr.github.io/BERTopic/api/bertopic.html}{berttopic}
}
}
\examples{
import_bertopic()
data <- sklearn::sk_datasets()
docs_all <-  data$fetch_20newsgroups(subset = 'all', remove = reticulate::tuple('headers', 'footers', 'quotes'))
docs <- docs_all["data"]
tm <- bert_topic()
topic_model <- tm$fit_transform(documents = docs)


}
