% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topic_documents}
\alias{bert_topic_documents}
\title{Bertopic Documents}
\usage{
bert_topic_documents(obj, docs, document_name = NULL)
}
\arguments{
\item{document_name}{}
}
\description{
Bertopic Documents
}
