% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-documents.R
\name{bert_topic_per_class}
\alias{bert_topic_per_class}
\title{Topics Per Structured Class}
\usage{
bert_topic_per_class(obj, docs = NULL, classes = NULL, global_tuning = TRUE)
}
\arguments{
\item{obj}{BERTopic model object.}

\item{docs}{Character vector of documents.}

\item{classes}{Character vector of class assignments, one per document.}

\item{global_tuning}{Logical. Fine-tune each topic representation for class by averaging its c-TF-IDF matrix with the global c-TF-IDF matrix. Default `TRUE`.}
}
\value{
A tibble with topic assignments per class, including `topic_bert`, `label_bertopic`, `class`, and word representations.
}
\description{
Topics Per Structured Class
}
