% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topics_keywords}
\alias{bert_topics_keywords}
\title{Returns keywords for the topics}
\usage{
bert_topics_keywords(topic_model)
}
\arguments{
\item{topic_model}{}
}
\description{
Returns keywords for the topics
}
