% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topics_over_time}
\alias{bert_topics_over_time}
\title{Extract Topics Over Time}
\usage{
bert_topics_over_time(
  obj,
  docs = NULL,
  timestamps = NULL,
  nr_bins = NULL,
  datetime_format = NULL,
  evolution_tuning = TRUE,
  global_tuning = TRUE,
  return_tibble = TRUE
)
}
\arguments{
\item{obj}{BERTopic model object.}

\item{docs}{Character vector of documents.}

\item{timestamps}{Vector of timestamps corresponding to documents.}

\item{nr_bins}{Integer. Number of time bins. Default NULL.}

\item{datetime_format}{Character. Format string for datetime. Default NULL.}

\item{evolution_tuning}{Logical. Enable evolution tuning. Default TRUE.}

\item{global_tuning}{Logical. Enable global tuning. Default TRUE.}

\item{return_tibble}{Logical. Return as tibble. Default TRUE.}
}
\value{
Topics over time data, as tibble if return_tibble is TRUE.
}
\description{
Analyze how topics change over time using timestamps.
}
\examples{
\dontrun{
topics_time <- bert_topics_over_time(tm, docs, timestamps)
}
}
