% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_transform_documents}
\alias{bert_transform_documents}
\title{Transform New Documents}
\usage{
bert_transform_documents(obj, documents = NULL, embeddings = NULL)
}
\arguments{
\item{obj}{BERTopic model object.}

\item{documents}{Character vector of documents to transform.}

\item{embeddings}{If not NULL, matrix or dataframe of pre-computed embeddings.}
}
\value{
A tibble with topic assignments and info for each document.
}
\description{
Transform new documents using an existing BERTopic model to predict topics.
}
\examples{
\dontrun{
new_docs <- c("New document about machine learning")
results <- bert_transform_documents(obj = tm, documents = new_docs)
}
}
