% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{bertopic_available}
\alias{bertopic_available}
\title{Is Python + BERTopic available?}
\usage{
bertopic_available()
}
\value{
Logical scalar.
}
\description{
Checks whether the active Python (as initialized by \pkg{reticulate}) can
import the key modules needed for BERTopic.
}
\examples{
\dontrun{
bertopic_available()
}
}
