% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{bertopic_get_document_info}
\alias{bertopic_get_document_info}
\title{Document-level information}
\usage{
bertopic_get_document_info(model, docs)
}
\arguments{
\item{model}{A "bertopic_r" model.}

\item{docs}{Character vector of documents to query (required).}
}
\value{
A tibble with document-level information.
}
\description{
Retrieve document-level information for the provided documents.
}
