% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{bertopic_has_embedding_model}
\alias{bertopic_has_embedding_model}
\title{Does the model have a usable embedding model?}
\usage{
bertopic_has_embedding_model(model)
}
\arguments{
\item{model}{A "bertopic_r" model.}
}
\value{
Logical; TRUE if \code{embedding_model} is present and not None.
}
\description{
Does the model have a usable embedding model?
}
