% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{bertopic_load}
\alias{bertopic_load}
\title{Load a BERTopic model from disk}
\usage{
bertopic_load(path)
}
\arguments{
\item{path}{Directory path passed to \code{\link[=bertopic_save]{bertopic_save()}}.}
}
\value{
A "bertopic_r" object with loaded Python model.
}
\description{
Load a BERTopic model from disk
}
