% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics_ops.R
\name{bertopic_reduce_topics}
\alias{bertopic_reduce_topics}
\title{Reduce/merge topics}
\usage{
bertopic_reduce_topics(
  model,
  nr_topics = "auto",
  representation_model = NULL,
  docs = NULL
)
}
\arguments{
\item{model}{A "bertopic_r" model.}

\item{nr_topics}{Target number (integer) or "auto".}

\item{representation_model}{Optional Python representation model.}

\item{docs}{Character vector of training docs.}
}
\value{
The input model (invisibly).
}
\description{
Wrapper over Python \code{reduce_topics}.
}
