% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{bertopic_session_info}
\alias{bertopic_session_info}
\title{Summarize Python/BERTopic session info}
\usage{
bertopic_session_info()
}
\value{
A named list containing paths, versions, and module availability:
\describe{
\item{python}{Path of the active Python.}
\item{libpython}{Path to libpython, if any.}
\item{version}{Python version string.}
\item{numpy}{Whether NumPy is available.}
\item{numpy_version}{NumPy version string (if available).}
\item{bertopic_version}{BERTopic version string (if available).}
\item{modules}{A data.frame with availability and exact installed versions for key modules.}
}
}
\description{
Summarize Python/BERTopic session info
}
\examples{
\dontrun{
bertopic_session_info()
}
}
