% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_terms.R
\name{bertopic_topic_terms}
\alias{bertopic_topic_terms}
\title{Get top terms for a topic}
\usage{
bertopic_topic_terms(model, topic_id, top_n = 10L)
}
\arguments{
\item{model}{A "bertopic_r" model}

\item{topic_id}{Integer topic id}

\item{top_n}{Number of top terms to return}
}
\value{
A tibble with columns \code{term} and \code{weight}
}
\description{
Get top terms for a topic
}
