% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{bertopic_topics}
\alias{bertopic_topics}
\title{Get topic info as a tibble}
\usage{
bertopic_topics(model)
}
\arguments{
\item{model}{A "bertopic_r" object returned by \code{\link[=bertopic_fit]{bertopic_fit()}}.}
}
\value{
A tibble with topic-level information from Python \code{get_topic_info()}.
}
\description{
Get topic info as a tibble
}
