% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{bertopic_visualize_barchart}
\alias{bertopic_visualize_barchart}
\title{Visualize a topic barchart}
\usage{
bertopic_visualize_barchart(model, topic_id = NULL, file = NULL)
}
\arguments{
\item{model}{A "bertopic_r" model.}

\item{topic_id}{Integer topic id. If NULL, a set of top topics is shown.}

\item{file}{Optional HTML output path.}
}
\description{
Visualize a topic barchart
}
