% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{bertopic_visualize_documents}
\alias{bertopic_visualize_documents}
\title{Visualize embedded documents}
\usage{
bertopic_visualize_documents(model, docs = NULL, file = NULL)
}
\arguments{
\item{model}{A "bertopic_r" model.}

\item{docs}{Optional character vector of documents to visualize.}

\item{file}{Optional HTML output path.}
}
\value{
An html file.
}
\description{
Visualize embedded documents
}
