% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{bertopic_visualize_heatmap}
\alias{bertopic_visualize_heatmap}
\title{Visualize topic similarity heatmap}
\usage{
bertopic_visualize_heatmap(model, file = NULL)
}
\arguments{
\item{model}{A "bertopic_r" model.}

\item{file}{Optional HTML output path.}
}
\description{
Visualize topic similarity heatmap
}
