% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{bertopic_visualize_hierarchy}
\alias{bertopic_visualize_hierarchy}
\title{Visualize hierarchical clustering of topics}
\usage{
bertopic_visualize_hierarchy(model, file = NULL)
}
\arguments{
\item{model}{A "bertopic_r" model.}

\item{file}{Optional HTML output path.}
}
\value{
An html file output.
}
\description{
Visualize hierarchical clustering of topics
}
