% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{bertopic_visualize_term_rank}
\alias{bertopic_visualize_term_rank}
\title{Visualize term rank evolution}
\usage{
bertopic_visualize_term_rank(model, file = NULL)
}
\arguments{
\item{model}{A "bertopic_r" model.}

\item{file}{Optional HTML output path.}
}
\description{
Visualize term rank evolution
}
