% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{bertopic_visualize_topics}
\alias{bertopic_visualize_topics}
\title{Visualize topic map}
\usage{
bertopic_visualize_topics(model, file = NULL)
}
\arguments{
\item{model}{A "bertopic_r" model.}

\item{file}{Optional HTML output path. If NULL, returns htmltools::HTML.}
}
\description{
Visualize topic map
}
