% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{bertopic_visualize_topics_over_time}
\alias{bertopic_visualize_topics_over_time}
\title{Visualize topics over time}
\usage{
bertopic_visualize_topics_over_time(
  model,
  topics_over_time,
  top_n = 10L,
  file = NULL
)
}
\arguments{
\item{model}{A "bertopic_r" model.}

\item{topics_over_time}{A tibble returned by \code{\link[=bertopic_topics_over_time]{bertopic_topics_over_time()}},
or a Python dataframe compatible with \code{visualize_topics_over_time()}.}

\item{top_n}{Number of topics to display.}

\item{file}{Optional HTML output path.}
}
\description{
Visualize topics over time
}
