% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{coef.bertopic_r}
\alias{coef.bertopic_r}
\title{Coefficients (top terms) for BERTopic}
\usage{
\method{coef}{bertopic_r}(object, top_n = 10L, ...)
}
\arguments{
\item{object}{A "bertopic_r" model.}

\item{top_n}{Number of terms per topic.}

\item{...}{Unused.}
}
\value{
A data.frame with columns topic, term, weight.
}
\description{
Coefficients (top terms) for BERTopic
}
