% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-io.R
\name{dictionary_nltk_stopwords}
\alias{dictionary_nltk_stopwords}
\title{NLTK Stopwords}
\usage{
dictionary_nltk_stopwords(language = "english")
}
\arguments{
\item{language}{Character. Language code. Default `"english"`.}
}
\value{
A character vector of stopwords in the specified language.
}
\description{
NLTK Stopwords
}
\examples{
dictionary_nltk_stopwords()
}
