% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{distilbert_embedding_transformer}
\alias{distilbert_embedding_transformer}
\title{Distilbert Transformer}
\usage{
distilbert_embedding_transformer(obj = NULL, ...)
}
\arguments{
\item{obj}{Transformers module object or `NULL`. If `NULL`, imports the module.}

\item{...}{Additional arguments passed to `pipeline()`.}
}
\value{
A Python feature-extraction pipeline using DistilBERT model.
}
\description{
Distilbert Transformer
}
\examples{
\dontrun{
pipeline <- distilbert_embedding_transformer()
}
}
