% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{flair_embeddings}
\alias{flair_embeddings}
\title{Access Flair Embeddings}
\usage{
flair_embeddings()
}
\value{
The Flair embeddings module object.
}
\description{
Access Flair Embeddings
}
\examples{
\dontrun{
embeddings <- flair_embeddings()
}
}
