% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{fortify.bertopic_r}
\alias{fortify.bertopic_r}
\title{Fortify method for ggplot2}
\usage{
\method{fortify}{bertopic_r}(model, data, ...)
}
\arguments{
\item{model}{A "bertopic_r" model.}

\item{data}{Ignored.}

\item{...}{Unused.}
}
\value{
A data.frame of document-topic assignments.
}
\description{
Fortify method for ggplot2
}
