% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-python.R
\name{get_py_env}
\alias{get_py_env}
\title{Return the Python environment name used by BERTopic}
\usage{
get_py_env()
}
\description{
Return the Python environment name used by BERTopic
}
\keyword{internal}
