% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{import_bertopic}
\alias{import_bertopic}
\title{Import Bert Topic}
\usage{
import_bertopic(
  assign_to_environment = T,
  path = NULL,
  numba_threads = 1,
  use_token_parallel = TRUE,
  use_kmp_fork = FALSE
)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE` assigns the BERTopic module to the global environment.}

\item{path}{Character. Path to Python environment if not using the default.}

\item{numba_threads}{Integer. Number of threads for Numba operations. Default `1`.}

\item{use_token_parallel}{Logical. If `TRUE` enables parallel tokenization. Default `TRUE`.}

\item{use_kmp_fork}{Logical. If `TRUE` uses KMP fork. Default `FALSE`.}
}
\value{
A Python BERTopic module object.
}
\description{
\href{https://maartengr.github.io/BERTopic}{BERTopic API from python} and other examples \href{https://github.com/MaartenGr/BERTopic}{here}
}
\examples{
library(bertopic)
import_bertopic()
}
