% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modules.R
\name{import_hdbscan}
\alias{import_hdbscan}
\title{Import hdbscan}
\usage{
import_hdbscan(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns module to global environment.}

\item{path}{Character. Path to Python installation directory.}
}
\value{
The Python hdbscan module object, invisibly.
}
\description{
Import hdbscan
}
\examples{
\dontrun{
hdbscan <- import_hdbscan(assign_to_environment = FALSE)
}
}
