% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{import_jellyfish}
\alias{import_jellyfish}
\title{Import Jellyfish}
\usage{
import_jellyfish(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE` assigns the Jellyfish module to the global environment.}

\item{path}{Character. Path to Python environment or `NULL` for default.}
}
\value{
A Python Jellyfish module object for string comparison.
}
\description{
Jelly fish module \url{https://jamesturk.github.io/jellyfish/}
}
