% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{import_keybert}
\alias{import_keybert}
\title{Import Keybert Module}
\usage{
import_keybert(
  assign_to_environment = T,
  use_token_parallel = TRUE,
  path = NULL
)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE` assigns to environment.}

\item{use_token_parallel}{Logical. If TRUE, sets TOKENIZERS_PARALLELISM=true env var. Default TRUE.}

\item{path}{Character. Python path to use for imports.}
}
\value{
A Python KeyBERT module object.
}
\description{
Import Keybert Module
}
\examples{
library(bertopic)
import_keybert()
}
