% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modules.R
\name{import_llama_cpp}
\alias{import_llama_cpp}
\title{Import Llama CPP}
\usage{
import_llama_cpp(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns module to global environment.}

\item{path}{Character. Path to Python installation directory.}
}
\value{
The Python llama-cpp module object, invisibly.
}
\description{
Import Llama CPP
}
\examples{
\dontrun{
llama_cpp <- import_llama_cpp(assign_to_environment = FALSE)
}
}
