% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modules.R
\name{import_nltk}
\alias{import_nltk}
\title{Import NLTK}
\usage{
import_nltk(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns module to global environment.}

\item{path}{Character. Path to Python installation directory.}
}
\value{
The Python NLTK module object, invisibly.
}
\description{
Import NLTK
}
\examples{
\dontrun{
nltk <- import_nltk(assign_to_environment = FALSE)
}
}
