% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modules.R
\name{import_nltk_corpus}
\alias{import_nltk_corpus}
\title{Import NLTK Corpus}
\usage{
import_nltk_corpus(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns module to global environment.}

\item{path}{Character. Path to Python installation directory.}
}
\value{
The Python NLTK corpus module object, invisibly.
}
\description{
Import NLTK Corpus
}
\examples{
\dontrun{
nltk_corpus <- import_nltk_corpus(assign_to_environment = FALSE)
}
}
