% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{import_numba}
\alias{import_numba}
\title{Import Numba}
\usage{
import_numba(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE` assigns the Numba module to the global environment.}

\item{path}{Character. Path to Python environment if not using the default.}
}
\value{
A Python Numba module object.
}
\description{
Import Numba
}
