% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{import_polyfuzz}
\alias{import_polyfuzz}
\title{Import Polyfuzz
Polyfuzz module \url{https://maartengr.github.io/PolyFuzz/}}
\usage{
import_polyfuzz(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE` assigns the PolyFuzz module to the global environment.}

\item{path}{Character. Path to Python environment or `NULL` for default.}
}
\value{
A Python PolyFuzz module object.
}
\description{
Import Polyfuzz
Polyfuzz module \url{https://maartengr.github.io/PolyFuzz/}
}
