% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{import_rapidfuzz}
\alias{import_rapidfuzz}
\title{Import RapidFuzz Module}
\usage{
import_rapidfuzz(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE` assigns the RapidFuzz module to the global environment.}

\item{path}{Character. Path to Python environment or `NULL` for default.}
}
\value{
A Python RapidFuzz module object.
}
\description{
RapidFuzz module \url{https://maxbachmann.github.io/RapidFuzz/}
}
