% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{import_sentence_transformers}
\alias{import_sentence_transformers}
\title{Sentence Transformer Module}
\usage{
import_sentence_transformers(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns module to global environment.}

\item{path}{Character. Path to Python installation directory.}
}
\value{
The Python sentence-transformers module object, invisibly.
}
\description{
Sentence Transformer Module
}
\examples{
\dontrun{
st <- import_sentence_transformers(assign_to_environment = FALSE)
}
}
