% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modules.R
\name{import_sklearn}
\alias{import_sklearn}
\title{Import Scikit Learn}
\usage{
import_sklearn(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns module to global environment.}

\item{path}{Character. Path to Python installation directory.}
}
\value{
The Python scikit-learn module object, invisibly.
}
\description{
Import Scikit Learn
}
\examples{
\dontrun{
sklearn <- import_sklearn(assign_to_environment = FALSE)
}
}
