% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modules.R
\name{import_spacy}
\alias{import_spacy}
\title{Import Spacy}
\usage{
import_spacy(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns module to global environment.}

\item{path}{Character. Path to Python installation directory.}
}
\value{
The Python Spacy module object, invisibly.
}
\description{
Import Spacy
}
\examples{
\dontrun{
spacy <- import_spacy(assign_to_environment = FALSE)
}
}
