% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top2vec.R
\name{import_top2vec}
\alias{import_top2vec}
\title{Import Top2Vec}
\usage{
import_top2vec(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns the imported Top2Vec module to the global environment. Default `TRUE`.}

\item{path}{Character. Optional path to Python executable. If `NULL`, uses default Python.}
}
\value{
A Python Top2Vec module object.
}
\description{
Import Top2Vec
}
