% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{import_transformers}
\alias{import_transformers}
\title{Import Hugging Face Transformers}
\usage{
import_transformers(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns module to global environment.}

\item{path}{Character. Path to Python installation directory.}
}
\value{
The Python transformers module object, invisibly.
}
\description{
Import Hugging Face Transformers
}
\examples{
\dontrun{
transformers <- import_transformers(assign_to_environment = FALSE)
}
}
