% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modules.R
\name{import_umap}
\alias{import_umap}
\title{Import UMAP}
\usage{
import_umap(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns module to global environment.}

\item{path}{Character. Path to Python installation directory.}
}
\value{
The Python UMAP module object, invisibly.
}
\description{
Import UMAP
}
\examples{
\dontrun{
umap <- import_umap(assign_to_environment = FALSE)
}
}
