% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modules.R
\name{import_use}
\alias{import_use}
\title{Import use}
\usage{
import_use(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{Logical. If `TRUE`, assigns module to global environment.}

\item{path}{Character. Path to Python installation directory.}
}
\value{
The Python use module object, invisibly.
}
\description{
Import use
}
\examples{
\dontrun{
use <- import_use(assign_to_environment = FALSE)
}
}
