% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{keybert_inspired_representation}
\alias{keybert_inspired_representation}
\title{Use a KeyBERT-like model to fine-tune the topic representations}
\usage{
keybert_inspired_representation(
  top_n_words = 10,
  nr_repr_docs = 5,
  nr_samples = 500,
  nr_candidate_words = 10,
  random_state = 42,
  numba_threads = 1,
  obj = NULL
)
}
\arguments{
\item{top_n_words}{The top n words to extract per topic.  Default `10`.}

\item{nr_repr_docs}{The number of representative documents to extract per cluster.  Default `5`.}

\item{nr_samples}{The number of candidate documents to extract per cluster.  Default `500`.}

\item{nr_candidate_words}{The number of candidate words per cluster.  Default `100`.}

\item{random_state}{The random state for randomly sampling candidate documents.  Default `42`.}

\item{obj}{Bertopic Module}
}
\description{
Use a KeyBERT-like model to fine-tune the topic representations
}
