% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{keybert_model}
\alias{keybert_model}
\title{Initiate Keybert model}
\usage{
keybert_model(keybert = NULL, model = "all-MiniLM-L6-v2")
}
\arguments{
\item{keybert}{Python KeyBERT module. If `NULL`, imports KeyBERT automatically.}

\item{model}{Character. Sentence transformer model to use. Default is `all-MiniLM-L6-v2`. See \href{https://www.sbert.net/docs/pretrained_models.html}{sbert documentation} for available models.}
}
\value{
A Python KeyBERT model object.
}
\description{
Runs a keybert model.
}
\examples{
kb <- keybert_model()
}
