% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyphrase.R
\name{keyphrase_vectorizer}
\alias{keyphrase_vectorizer}
\title{Build a Keyphrase Vectorizer}
\usage{
keyphrase_vectorizer(
  obj = NULL,
  language = "english",
  stopword_package_sources = NULL,
  exclude_stop_words = T,
  extra_stop_words = NULL,
  is_lower_case = TRUE,
  spacy_exclude = list("parser", "attribute_ruler", "lemmatizer", "ner"),
  max_df = NULL,
  min_df = NULL,
  workers = 1,
  pos_pattern = "<J.*>*<N.*>+",
  spacy_pipeline = "en_core_web_sm",
  custom_pos_tagger = NULL,
  binary = FALSE,
  decay = NULL,
  delete_min_df = NULL
)
}
\arguments{
\item{obj}{Keyphrase module object. If `NULL`, imports automatically.}

\item{language}{Character. Stopwords language. Default `"english"`.}

\item{stopword_package_sources}{Character vector. External stopword sources. Default `NULL`.}

\item{exclude_stop_words}{Logical. If `TRUE`, excludes stopwords. Default `TRUE`.}

\item{extra_stop_words}{Character vector. Additional stopwords to exclude. Default `NULL`.}

\item{is_lower_case}{Logical. If `TRUE`, converts keyphrases to lowercase. Default `TRUE`.}

\item{spacy_exclude}{List. spaCy pipeline components to exclude. Default `list('parser', 'attribute_ruler', 'lemmatizer', 'ner')`.}

\item{max_df}{Integer or `NULL`. Maximum document frequency threshold. Default `NULL`.}

\item{min_df}{Integer or `NULL`. Minimum document frequency threshold. Default `NULL`.}

\item{workers}{Integer. Number of spaCy workers. Default `1`.}

\item{pos_pattern}{Character. POS pattern for keyphrase extraction. Default `"<J.*>*<N.*>+"`.}

\item{spacy_pipeline}{Character. spaCy model. Default `"en_core_web_sm"`.}

\item{custom_pos_tagger}{Callable or `NULL`. Custom POS tagger function. Default `NULL`.}

\item{binary}{Logical. If `TRUE`, all non-zero counts are set to 1. Default `FALSE`.}

\item{decay}{Numeric or `NULL`. Decay parameter. Default `NULL`.}

\item{delete_min_df}{Logical or `NULL`. Delete min_df parameter. Default `NULL`.}
}
\value{
A Python KeyphraseCountVectorizer object configured with specified parameters.
}
\description{
Build a Keyphrase Vectorizer
}
\examples{
vectorizer_model <- keyphrase_vectorizer(pos_pattern = "<J.*>*<N.*>+")
docs <- c('This is the first document.', 'This document is the second document.', 'And this is the third one.', 'Is this the first document?')
vectorizer_model$fit_transform(raw_documents = docs)
vectorizer_model$get_feature_names_out()
vectorizer_model$keyphrases
}
