% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-analysis.R
\name{list_to_tuple}
\alias{list_to_tuple}
\title{List to Tuple}
\usage{
list_to_tuple(obj)
}
\arguments{
\item{obj}{List object to convert to a Python tuple.}
}
\value{
A Python tuple object.
}
\description{
List to Tuple
}
