% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-analysis.R
\name{munge_bert_topics_over_time}
\alias{munge_bert_topics_over_time}
\title{Munge Topic Over Time}
\usage{
munge_bert_topics_over_time(data, topic_model = NULL)
}
\arguments{
\item{data}{Output from bert_topics_over_time function.}

\item{topic_model}{BERTopic model object to add topic labels. Default `NULL`.}
}
\value{
A tibble with cleaned topics over time data, with columns: `topic_bert`, `words`, `count`, `date_time`, `date`, and optionally topic labels from the model.
}
\description{
Transform topics over time output into a tidy tibble format.
}
\examples{
\dontrun{
tidy_time <- munge_bert_topics_over_time(topics_time_data, tm)
}
}
