% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{part_of_speech_representation}
\alias{part_of_speech_representation}
\title{Extract Topic Keywords based on their Part-of-Speech}
\usage{
part_of_speech_representation(
  model = "en_core_web_sm",
  top_n_words = 10,
  pos_patterns = NULL,
  obj = NULL
)
}
\arguments{
\item{model}{Character. The Spacy model to use. Default `"en_core_web_sm"`.}

\item{top_n_words}{Integer. The top n words to extract. Default `10`.}

\item{pos_patterns}{List or `NULL`. Patterns for Spacy to use for matching. See Spacy documentation for pattern structure.}

\item{obj}{BERTopic object or `NULL`.}
}
\value{
A Python PartOfSpeech representation object with representation_method and input_parameters attributes.
}
\description{
Extract Topic Keywords based on their Part-of-Speech
}
\examples{
\dontrun{
library(bertopic)
part_of_speech_representation(
  pos_patterns = list(
    list(list('POS' = 'ADJ'), list('POS' = 'NOUN')),
    list(list('POS' = 'NOUN')),
    list(list('POS' = 'ADJ'))
  )
)
}

}
