% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_edit_distance}
\alias{polyfuzz_edit_distance}
\title{Calculate the Edit Distance between lists of strings using any distance/similarity based scorer}
\usage{
polyfuzz_edit_distance(
  n_jobs = 1,
  scorer = NULL,
  model_id = NULL,
  normalize = TRUE
)
}
\arguments{
\item{n_jobs}{Integer. Number of parallel jobs. Default is 1.}

\item{scorer}{A Python function for calculating edit distance. Should return a float between 0 and 1. If `NULL` uses jellyfish `jaro_winkler_similarity`.}

\item{model_id}{Character. Name of the particular instance, used when comparing models.}

\item{normalize}{Logical. Whether to normalize scores. Default is `TRUE`.}
}
\value{
A Python PolyFuzz EditDistance model object.
}
\description{
Calculate the Edit Distance between lists of strings using any distance/similarity based scorer
}
