% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_embeddings}
\alias{polyfuzz_embeddings}
\title{Embedding Method}
\usage{
polyfuzz_embeddings(
  embedding_method = NULL,
  min_similarity = 0.75,
  top_n = 1,
  cosine_method = "sparse",
  model_id = NULL
)
}
\arguments{
\item{embedding_method}{A Flair embeddings object or `NULL` for default.}

\item{min_similarity}{Numeric. Minimum similarity between strings, otherwise return 0. Default is 0.75.}

\item{top_n}{Integer. The number of best matches to return.}

\item{cosine_method}{Character. Method for calculating cosine similarity. Options: `sparse`, `sklearn`, `knn`. Default is `sparse`.}

\item{model_id}{Character. Name of the model instance.}
}
\value{
A Python PolyFuzz Embeddings model object.
}
\description{
Embedding Method
}
