% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_model}
\alias{polyfuzz_model}
\title{Polyfuzz Model}
\usage{
polyfuzz_model(
  method = "TF-IDF",
  model_object = NULL,
  obj = NULL,
  verbose = TRUE,
  from_list = NULL,
  to_list = NULL
)
}
\arguments{
\item{method}{Character. The method used for matching. Options: `EditDistance`, `TF-IDF`, or `Embeddings`. Default is `TF-IDF`.}

\item{model_object}{A pretrained model object from polyfuzz or `NULL` for default.}

\item{obj}{A PolyFuzz module object or `NULL` for default import.}

\item{verbose}{Logical. If `TRUE` prints results during fitting.}

\item{from_list}{Character vector. Words to match from, or `NULL`.}

\item{to_list}{Character vector. Words to match to, or `NULL`.}
}
\value{
A Python PolyFuzz model object.
}
\description{
Polyfuzz Model
}
