% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_rapidmatcher}
\alias{polyfuzz_rapidmatcher}
\title{Calculate the Edit Distance between lists of strings using RapidFuzz's process function}
\usage{
polyfuzz_rapidmatcher(
  n_jobs = 1,
  score_cutoff = 0.5,
  scorer = NULL,
  model_id = NULL
)
}
\arguments{
\item{n_jobs}{Integer. Number of parallel processes; use -1 for all cores. Default is 1.}

\item{score_cutoff}{Numeric. Minimum similarity for a good match, between 0 and 1. Default is 0.5.}

\item{scorer}{A Python scorer function. Options: `fuzz.ratio`, `fuzz.partial_ratio`, `fuzz.token_sort_ratio`, `fuzz.partial_token_sort_ratio`, `fuzz.token_set_ratio`, `fuzz.partial_token_set_ratio`, `fuzz.token_ratio`, `fuzz.partial_token_ratio`, `fuzz.WRatio`, `fuzz.QRatio`. If `NULL` uses WRatio.}

\item{model_id}{Character. Name of the particular instance, used when comparing models.}
}
\value{
A Python PolyFuzz RapidFuzz model object.
}
\description{
We are using RapidFuzz instead of FuzzyWuzzy since it is much faster and does not require the more restrictive GPL license
}
