% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_save}
\alias{polyfuzz_save}
\title{Save Polyfuzz Model}
\usage{
polyfuzz_save(obj, model_path = NULL, file_name = "polyfuzz")
}
\arguments{
\item{obj}{A PolyFuzz model object.}

\item{model_path}{Character. Path to directory where model will be saved.}

\item{file_name}{Character. Name of the file. Defaults to polyfuzz.}
}
\value{
Called for side effects; returns invisibly. Saves model to specified path.
}
\description{
Save Polyfuzz Model
}
