% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_sentence_embeddings}
\alias{polyfuzz_sentence_embeddings}
\title{Embed words into vectors and use cosine similarity to find
the best matches between two lists of strings}
\usage{
polyfuzz_sentence_embeddings(
  embedding_model = NULL,
  min_similarity = 0.5,
  top_n = 1,
  cosine_method = "sparse",
  model_id = NULL
)
}
\arguments{
\item{embedding_model}{A Sentence Transformer model. Can be either a string or model object.}

\item{min_similarity}{Numeric. Minimum similarity between strings, otherwise return 0. Default is 0.5.}

\item{top_n}{Integer. The number of best matches to return.}

\item{cosine_method}{Character. Method for calculating cosine similarity. Options: `sparse`, `sklearn`, `knn`. Default is `sparse`.}

\item{model_id}{Character. Name of the model instance.}
}
\value{
A Python PolyFuzz SentenceEmbeddings model object.
}
\description{
Embed words into vectors and use cosine similarity to find
the best matches between two lists of strings
}
