% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_tfidf}
\alias{polyfuzz_tfidf}
\title{A character based n-gram TF-IDF to approximate edit distance
We turn a string into, typically of length 3, n-grams. For example, using 3-grams of the "hotel" we get ['hot', 'ote', 'tel']. These are then used as input for a TfidfVectorizer in order to create a vector for each word. Then, we simply apply cosine similarity through k-NN}
\usage{
polyfuzz_tfidf(
  n_gram_range = list(3L, 3L),
  clean_string = TRUE,
  min_similarity = 0.75,
  top_n = 1,
  cosine_method = "sparse",
  model_id = NULL
)
}
\arguments{
\item{n_gram_range}{List of integers. The n-gram range on a character-level. Default is `list(3L, 3L)`.}

\item{clean_string}{Logical. If `TRUE` cleans the string to keep only alphanumeric characters. Default is `TRUE`.}

\item{min_similarity}{Numeric. Minimum similarity between strings, otherwise return 0. Default is 0.75.}

\item{top_n}{Integer. The number of matches you want returned.}

\item{cosine_method}{Character. Method for calculating cosine similarity. Options: `sparse`, `sklearn`, `knn`. Default is `sparse`.}

\item{model_id}{Character. Name of the particular instance, used when comparing models.}
}
\value{
A Python PolyFuzz TFIDF model object.
}
\description{
A character based n-gram TF-IDF to approximate edit distance
We turn a string into, typically of length 3, n-grams. For example, using 3-grams of the "hotel" we get ['hot', 'ote', 'tel']. These are then used as input for a TfidfVectorizer in order to create a vector for each word. Then, we simply apply cosine similarity through k-NN
}
