% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_use_embeddings}
\alias{polyfuzz_use_embeddings}
\title{USE model}
\usage{
polyfuzz_use_embeddings(
  embedding_model = "https://tfhub.dev/google/universal-sentence-encoder/4",
  min_similarity = 0.5,
  top_n = 1,
  cosine_method = "sparse",
  model_id = NULL
)
}
\arguments{
\item{embedding_model}{Character. URL to the Universal Sentence Encoder model. Default is `"https://tfhub.dev/google/universal-sentence-encoder/4"`.}

\item{min_similarity}{Numeric. Minimum similarity between strings, otherwise return 0. Default is 0.5.}

\item{top_n}{Integer. The number of best matches to return.}

\item{cosine_method}{Character. Method for calculating cosine similarity. Options: `sparse`, `sklearn`, `knn`. Default is `sparse`.}

\item{model_id}{Character. Name of the model instance.}
}
\value{
A Python PolyFuzz USEEmbeddings model object.
}
\description{
USE model
}
