% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporters.R
\name{predict.bertopic_r}
\alias{predict.bertopic_r}
\title{Predict method for BERTopic models}
\usage{
\method{predict}{bertopic_r}(
  object,
  newdata,
  type = c("both", "class", "prob"),
  embeddings = NULL,
  ...
)
}
\arguments{
\item{object}{A "bertopic_r" model.}

\item{newdata}{Character vector of new documents.}

\item{type}{One of "class", "prob", or "both".}

\item{embeddings}{Optional numeric matrix of embeddings.}

\item{...}{Reserved for future arguments.}
}
\value{
Depending on \code{type}, an integer vector, a matrix/data frame, or a list.
}
\description{
Predict method for BERTopic models
}
