% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.bertopic_r}
\alias{print.bertopic_r}
\title{Print method for bertopic_r}
\usage{
\method{print}{bertopic_r}(x, ...)
}
\arguments{
\item{x}{A "bertopic_r" object.}

\item{...}{Unused.}
}
\description{
Print method for bertopic_r
}
