% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{roberta_embeddings}
\alias{roberta_embeddings}
\title{Roberta Embeddings}
\usage{
roberta_embeddings(obj = NULL, is_token_embedding = T)
}
\arguments{
\item{obj}{Flair module object or `NULL`. If `NULL`, imports the module.}

\item{is_token_embedding}{Logical. If `TRUE`, uses token-level embeddings. Default `TRUE`.}
}
\value{
A Python TransformerDocumentEmbeddings object for RoBERTa embeddings.
}
\description{
Roberta Embeddings
}
\examples{
\dontrun{
roberta <- roberta_embeddings()
}
}
