% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{set_bertopic_seed}
\alias{set_bertopic_seed}
\title{Set random seed for R and Python backends}
\usage{
set_bertopic_seed(seed)
}
\arguments{
\item{seed}{Integer seed}
}
\description{
Set random seed for R and Python backends
}
