% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{spacy_matching_pattern_dict_to_list}
\alias{spacy_matching_pattern_dict_to_list}
\title{Convert Spacy Matching Pattern to Dictionary}
\usage{
spacy_matching_pattern_dict_to_list(
  pattern = "[{'POS': 'ADJ'},\\n           {'POS': 'NOUN'}]"
)
}
\arguments{
\item{pattern}{Character. A Spacy matching pattern as a string to convert to R list format.}
}
\value{
A list representation of the Spacy matching pattern.
}
\description{
Convert Spacy Matching Pattern to Dictionary
}
\examples{
\dontrun{
pattern <- "[{'POS': 'ADJ'}, {'POS': 'NOUN'}]"
spacy_matching_pattern_dict_to_list(pattern = pattern)
}
}
