% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{summary.bertopic_r}
\alias{summary.bertopic_r}
\title{Summary for BERTopic models}
\usage{
\method{summary}{bertopic_r}(object, ...)
}
\arguments{
\item{object}{A "bertopic_r" model.}

\item{...}{Unused.}
}
\value{
Invisibly returns a named list of summary fields.
}
\description{
Summary for BERTopic models
}
