% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{tbl_bert_attributes}
\alias{tbl_bert_attributes}
\title{Return Tibble of BERT Attributes}
\usage{
tbl_bert_attributes(
  obj,
  return_clean = F,
  return_wide = F,
  parameter_filter = NULL
)
}
\arguments{
\item{obj}{A BERTopic model object.}

\item{return_clean}{Logical. If `TRUE` returns a clean set of parameters. Default `FALSE`.}

\item{return_wide}{Logical. If `TRUE` returns it in wide form. Default `FALSE`.}

\item{parameter_filter}{Character vector or `NULL`. If not `NULL` filters the attributes by pattern matching. Default `NULL`.}
}
\value{
A tibble containing model attributes, either in long or wide format depending on arguments.
}
\description{
Return Tibble of BERT Attributes
}
