% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{tbl_bert_fit_transform}
\alias{tbl_bert_fit_transform}
\title{BERTopic Fit and Transform from tibble.}
\usage{
tbl_bert_fit_transform(
  obj,
  data = NULL,
  text_field = NULL,
  class_fields = NULL,
  embeddings = NULL,
  images = NULL
)
}
\arguments{
\item{obj}{A BERTopic model object created by `bert_topic()`.}

\item{data}{A tibble containing text and optional class columns.}

\item{text_field}{Character. Name of the column containing text documents.}

\item{class_fields}{Character vector or `NULL`. Names of columns to use as class labels for semi-supervised learning. Default `NULL`.}

\item{embeddings}{Numeric matrix or `NULL`. Pre-trained document embeddings. These can be used instead of the sentence-transformer model. Default `NULL`.}

\item{images}{List or `NULL`. A list of paths to images or the images themselves. Default `NULL`.}
}
\value{
An integer vector of assigned topic IDs for each document.
}
\description{
Fit the models on a collection of documents, generate topics, and return the docs with topics.
}
