% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-documents.R
\name{tbl_bert_topic_per_class}
\alias{tbl_bert_topic_per_class}
\title{Find Topics from Class via Tibble}
\usage{
tbl_bert_topic_per_class(
  data,
  topic_model,
  document_name = NULL,
  class_name = NULL,
  global_tuning = TRUE,
  sort_by_topic = TRUE
)
}
\arguments{
\item{data}{A tibble or data frame containing documents and class assignments.}

\item{topic_model}{BERTopic model object.}

\item{document_name}{Character. Name of the text feature column.}

\item{class_name}{Character vector. Name(s) of the class feature column(s).}

\item{global_tuning}{Logical. Fine-tune each topic representation for class by averaging its c-TF-IDF matrix with the global c-TF-IDF matrix. Default `TRUE`.}

\item{sort_by_topic}{Logical. If TRUE, arranges output by label_bertopic descending then count descending. Default FALSE.}
}
\value{
A tibble with topic assignments per class, including `topic_bert`, `label_bertopic`, class columns, and word representations.
}
\description{
Find Topics from Class via Tibble
}
