% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-documents.R
\name{tbl_bert_transform_documents}
\alias{tbl_bert_transform_documents}
\title{Transform Documents from Tibble}
\usage{
tbl_bert_transform_documents(
  data,
  topic_model,
  document_name = NULL,
  embeddings = TRUE
)
}
\arguments{
\item{data}{A tibble or data frame containing documents to transform.}

\item{topic_model}{BERTopic model object.}

\item{document_name}{Character. Name of the column containing documents.}

\item{embeddings}{Logical or matrix. Pre-computed embeddings. If `TRUE`, computes embeddings. Default `TRUE`.}
}
\value{
A tibble containing original data with added topic assignments and related information.
}
\description{
Transform Documents from Tibble
}
