% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{tbl_keybert_data}
\alias{tbl_keybert_data}
\title{Extract Keybert Keywords from Output}
\usage{
tbl_keybert_data(out, use_future = FALSE, return_message = FALSE, slug = NULL)
}
\arguments{
\item{out}{Python KeyBERT output object containing extracted keywords.}

\item{use_future}{Logical. If `TRUE`, uses parallel processing. Default `FALSE`.}

\item{return_message}{Logical. If `TRUE`, prints processing messages. Default `FALSE`.}
}
\value{
A tibble with columns: `number_document`, `keyword_keybert`, and `score_keybert`.
}
\description{
Extract Keybert Keywords from Output
}
