% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top2vec.R
\name{top2vec_document_topics}
\alias{top2vec_document_topics}
\title{Top2Vec Documents}
\usage{
top2vec_document_topics(obj)
}
\arguments{
\item{obj}{Top2Vec model object.}
}
\value{
A tibble with columns: `number_document`, `id_document`, `topic_top2vec`, and `distance`.
}
\description{
Top2Vec Documents
}
