% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top2vec.R
\name{top2vec_topic_embeddings}
\alias{top2vec_topic_embeddings}
\title{Topic Vector Embeddings}
\usage{
top2vec_topic_embeddings(obj)
}
\arguments{
\item{obj}{Top2Vec model object.}
}
\value{
A tibble with topic vector embeddings, with `topic_top2vec` as the first column.
}
\description{
Topic Vector Embeddings
}
