% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{use_bertopic}
\alias{use_bertopic}
\title{Bind current R session to the BERTopic environment (auto route)}
\usage{
use_bertopic(envname = "r-bertopic")
}
\arguments{
\item{envname}{Character. Environment name. Default "r-bertopic".}
}
\value{
Invisibly, the Python executable path.
}
\description{
If a Conda env with the given name exists, prefer Conda; otherwise try a
virtualenv with the same name. Stops if neither exists.
}
