% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{base_embedder}
\alias{base_embedder}
\title{The Base Embedder used for creating embedding models}
\usage{
base_embedder(embedding_model = NULL, word_embedding_model = NULL)
}
\arguments{
\item{embedding_model}{The main embedding model to be used for extracting document and word embedding default `NULL`}

\item{word_embedding_model}{The embedding model used for extracting word embeddings only. If this model is selected, then the embedding_model is purely used for creating document embeddings. Default `NULL`}
}
\description{
The Base Embedder used for creating embedding models
}
