% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_document_info}
\alias{bert_document_info}
\title{Bertopic Documents}
\usage{
bert_document_info(obj, docs, document_name = NULL)
}
\arguments{
\item{obj}{Topic model object}

\item{docs}{vector of documents}

\item{document_name}{If not `NULL` name of the text document column}
}
\description{
Bertopic Documents
}
