% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{bert_embed_documents}
\alias{bert_embed_documents}
\title{Embed Documents}
\usage{
bert_embed_documents(obj, docs, return_tibble = TRUE, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\description{
Embed Documents
}
