% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{bert_embed_words}
\alias{bert_embed_words}
\title{Embed Vector of Words}
\usage{
bert_embed_words(obj, words, return_tibble = T, verbose = TRUE)
}
\arguments{
\item{obj}{Topic Model Object}

\item{words}{Vector of Words}

\item{verbose}{if `TRUE` verbose}
}
\description{
Embed Vector of Words
}
