% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_embeddings}
\alias{bert_embeddings}
\title{Bert Embeddings}
\usage{
bert_embeddings(embeddings, docs = NULL, id_columns = NULL, text_column = NULL)
}
\arguments{
\item{text_column}{}
}
\description{
Bert Embeddings
}
