% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_fit}
\alias{bert_fit}
\title{BERTopic Fit}
\usage{
bert_fit(obj, documents, embeddings = NULL, y = NULL, images = images)
}
\arguments{
\item{obj}{BERTopic Object}

\item{documents}{A single document or a list of documents to fit on}

\item{embeddings}{If not `NULL` Pre-trained document embeddings. These can be used instead of the sentence-transformer model.}

\item{y}{If Not `NULL` The target class for (semi)-supervised modeling. Use -1 if no class for a specific instance is specified.}
}
\description{
Fit the models (Bert, UMAP, and, HDBSCAN) on a collection of documents and generate topics
}
