% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_fit_transform}
\alias{bert_fit_transform}
\title{BERTopic Fit and Transform}
\usage{
bert_fit_transform(obj, documents, embeddings = NULL, y = NULL)
}
\arguments{
\item{obj}{BERTopic Object}

\item{documents}{A single document or a list of documents to fit on}

\item{embeddings}{If not `NULL` Pre-trained document embeddings. These can be used instead of the sentence-transformer model.}

\item{y}{If Not `NULL` The target class for (semi)-supervised modeling. Use -1 if no class for a specific instance is specified.#'}
}
\description{
Fit the models on a collection of documents, generate topics, and return the docs with topics.
}
