% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_load}
\alias{bert_load}
\title{Load BERTopic Model}
\usage{
bert_load(
  model_path = NULL,
  obj = NULL,
  numba_threads = 1,
  use_token_parallel = TRUE,
  embedding_model = NULL
)
}
\arguments{
\item{use_token_parallel}{}
}
\description{
Load BERTopic Model
}
