% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_merge_models}
\alias{bert_merge_models}
\title{Merge multiple pre-trained BERTopic models into a single model}
\usage{
bert_merge_models(
  obj,
  models = NULL,
  min_similarity = 0.7,
  embedding_model = NULL
)
}
\arguments{
\item{obj}{BERTopic object}

\item{models}{A list of fitted BERTopic models}

\item{min_similarity}{The minimum similarity for when topics are merged.  Default .7}

\item{embedding_model}{Additionally load in an embedding model if necessary.}
}
\description{
Merge multiple pre-trained BERTopic models into a single model
}
