% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_merge_topics}
\alias{bert_merge_topics}
\title{Merge Topics}
\usage{
bert_merge_topics(obj, docs, topics_to_merge = NULL)
}
\arguments{
\item{obj}{bertopic model object}

\item{docs}{The documents you used when calling either fit or fit_transform}

\item{topics_to_merge}{Either a list of topics or a list of list of topics to merge. For example: `list(1,2,3)` will merge topics 1, 2 and 3 `list(list(1,2), list(3,4))` will merge topics 1 and 2, and separately merge topics 3 and 4.}
}
\description{
Merge Topics
}
