% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_parameters}
\alias{bert_parameters}
\title{BERTopic Parameters}
\usage{
bert_parameters(obj, deep = TRUE)
}
\arguments{
\item{obj}{topic model object}

\item{deep}{if `TRUE` returns deep information}
}
\description{
BERTopic Parameters
}
