% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_parameters}
\alias{bert_parameters}
\title{BERTopic Parameters}
\usage{
bert_parameters(
  obj,
  deep = FALSE,
  return_tibble = FALSE,
  use_dfc_method = FALSE,
  return_attributes = FALSE
)
}
\arguments{
\item{obj}{topic model object}

\item{deep}{if `TRUE` returns deep information}

\item{return_tibble}{If `TRUE` returns a `tibble`}

\item{return_attributes}{If `TRUE` returns concatenated character vector of the parameters}
}
\description{
BERTopic Parameters
}
