% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_partial_fit}
\alias{bert_partial_fit}
\title{Fit BERTopic on a subset of the data and perform online learning with batch-like data.}
\usage{
bert_partial_fit(obj, documents, embeddings = NULL, y = NULL)
}
\arguments{
\item{obj}{Topic Model Object}

\item{documents}{A list of documents to fit on}

\item{embeddings}{Pre-trained document embeddings. These can be used instead of the sentence-transformer model}

\item{y}{The target class for (semi)-supervised modeling. Use -1 if no class for a specific instance is specified}
}
\description{
Fit BERTopic on a subset of the data and perform online learning with batch-like data.
}
