% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_reduce_outliers}
\alias{bert_reduce_outliers}
\title{Reduce BERTopics}
\usage{
bert_reduce_outliers()
}
\arguments{
\item{obj}{BERTopic Topic Model}

\item{docs}{Vector of Documents}

\item{number_topics}{Number of topics to reduce to}

\item{number_words}{Top n words per topic to use.  Default is 4}

\item{separator}{The string with which the words and topic prefix will be separated. Underscores are the default but a nice alternative is ", ".  Default `_`}

\item{word_length}{The maximum length of each word in the topic label. Some words might be relatively long and setting this value helps to make sure that all labels have relatively similar lengths.}

\item{topic_prefix}{Whether to use the topic ID as a prefix. If set to True, the topic ID will be separated using the separator}

\item{append_number_words}{Append the number of words}

\item{update_bert_labels}{If `TRUE` updates actual topic labels}

\item{update_topic_model_labels}{if `TRUE` updates new label into a custom field}
}
\description{
Reduce BERTopics
}
