% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_reduce_topics}
\alias{bert_reduce_topics}
\title{Reduce BERTopics}
\usage{
bert_reduce_topics(obj, docs = NULL, number_topics = NULL)
}
\arguments{
\item{obj}{BERTopic Topic Model}

\item{docs}{Vector of Documents}

\item{number_topics}{Number of topics to reduce to}
}
\description{
Reduce BERTopics
}
