% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_representative_documents}
\alias{bert_representative_documents}
\title{Topic's Representative Documents}
\usage{
bert_representative_documents(
  obj,
  topic_number = NULL,
  include_labels = T,
  number_words = 5L,
  top_n_documents = NULL,
  sep = "_"
)
}
\arguments{
\item{obj}{Topic Model Object}

\item{topic_number}{If not `NULL` number of topic}
}
\description{
Topic's Representative Documents
}
