% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_representative_documents}
\alias{bert_representative_documents}
\title{Topic's Representative Documents}
\usage{
bert_representative_documents(
  topic_model,
  topic_number = NULL,
  include_labels = T,
  label_words = 5L
)
}
\arguments{
\item{topic_number}{}
}
\description{
Topic's Representative Documents
}
