% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_save}
\alias{bert_save}
\title{Save BERTopic Model}
\usage{
bert_save(obj = NULL, file_path = NULL, file_name = "bert_model")
}
\arguments{
\item{save_embedding_model}{}
}
\description{
Save BERTopic Model
}
