% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_save}
\alias{bert_save}
\title{Save a Bertopic Model}
\usage{
bert_save(
  obj,
  file_path = NULL,
  file_name = "bert_model",
  save_embedding_model = TRUE
)
}
\arguments{
\item{save_embedding_model}{}
}
\description{
Save a Bertopic Model
}
