% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_similar_terms_topics}
\alias{bert_similar_terms_topics}
\title{Simllar Term Embeddings}
\usage{
bert_similar_terms_topics(
  topic_model,
  terms = NULL,
  top_n_terms = 10L,
  nest_data = F,
  return_message = T
)
}
\arguments{
\item{topic_model}{Bertopic Model}

\item{terms}{Vector of Terms to search}

\item{top_n_terms}{the number of topics to return.  Default is 10}

\item{nest_data}{if `TRUE` Nests data}
}
\description{
Simllar Term Embeddings
}
