% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_similar_terms_topics}
\alias{bert_similar_terms_topics}
\title{Simlar Term Embeddings}
\usage{
bert_similar_terms_topics(
  topic_model,
  terms = NULL,
  top_n_terms = 10L,
  nest_data = F,
  return_message = T
)
}
\arguments{
\item{nest_data}{}
}
\description{
Simlar Term Embeddings
}
