% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_similar_terms_topics}
\alias{bert_similar_terms_topics}
\title{Find topics most similar to Search Terms}
\usage{
bert_similar_terms_topics(
  obj,
  terms = NULL,
  top_n_terms = 10L,
  nest_data = F,
  image = NULL,
  return_message = T
)
}
\arguments{
\item{obj}{BERTopic Ojbect}

\item{terms}{Vector of Terms to search}

\item{top_n_terms}{the number of topics to return.  Default is 10}

\item{nest_data}{if `TRUE` Nests data}

\item{return_message}{}
}
\description{
Find topics most similar to Search Terms
}
