% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_stopwords}
\alias{bert_stopwords}
\title{Stopword List Generator}
\usage{
bert_stopwords(
  language = "english",
  is_lower_case = T,
  extra_stop_words = NULL
)
}
\arguments{
\item{extra_stop_words}{}
}
\description{
Stopword List Generator
}
