% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_topic}
\alias{bert_topic}
\title{Initiate BERT Topic Model}
\usage{
bert_topic()
}
\value{
python object
}
\description{
Functions for unsupervised clustering algorithms.
}
\details{

}
\examples{
import_bertopic()
data <- sklearn::sk_datasets()
docs <- data$fetch_20newsgroups(subset = 'all', remove = c('headers', 'footers', 'quotes'))
docs <- docs["data"]
tm <- bert_topic()
topic_model <- tm$fit_transform(documents = docs)


}
