% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topic_count}
\alias{bert_topic_count}
\title{Topic Counts}
\usage{
bert_topic_count(
  obj,
  topic_number = NULL,
  include_representative_documents = TRUE,
  join_labels = TRUE,
  include_parameters = FALSE,
  parameter_filters = NULL
)
}
\arguments{
\item{obj}{bertopic object}

\item{topic_number}{if not `NULL` the number of the topic.  If `NULL` returns all topics.}
}
\description{
Topic Counts
}
