% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topic_info}
\alias{bert_topic_info}
\title{BERT Topic Model Info}
\usage{
bert_topic_info(
  obj,
  topic_number = NULL,
  return_summary = TRUE,
  assign_to_environment = TRUE,
  concatenator = ".  ",
  top_n_aspect_words = NULL,
  remove_list_columns = TRUE
)
}
\arguments{
\item{top_n_aspect_words}{}
}
\description{
BERT Topic Model Info
}
