% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topic_info}
\alias{bert_topic_info}
\title{BERT Topic Model Info}
\usage{
bert_topic_info(
  obj,
  topic_number = NULL,
  assign_to_environment = TRUE,
  concatenator = ".  ",
  remove_list_columns = FALSE
)
}
\arguments{
\item{remove_list_columns}{}
}
\description{
BERT Topic Model Info
}
