% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topic_keywords}
\alias{bert_topic_keywords}
\title{Returns keywords for the topics}
\usage{
bert_topic_keywords(obj, bert_topics = NULL, return_full_data = FALSE)
}
\arguments{
\item{obj}{Topic Model Object}

\item{bert_topics}{if not `NULL` specific topics}

\item{return_full_data}{If True, returns all different forms of topic representations for each topic, including aspects}
}
\description{
Returns keywords for the topics
}
