% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topic_labels}
\alias{bert_topic_labels}
\title{Generate BERT Topic Labels}
\usage{
bert_topic_labels(
  topic_model,
  number_words = 4L,
  separator = "_",
  word_length = NULL
)
}
\arguments{
\item{word_length}{}
}
\description{
Generate BERT Topic Labels
}
