% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topic_labels}
\alias{bert_topic_labels}
\title{Generate BERT Topic Labels}
\usage{
bert_topic_labels(
  obj,
  number_words = 4L,
  separator = "_",
  word_length = NULL,
  update_topic_model_labels = FALSE,
  append_number_words = FALSE,
  topic_prefix = FALSE
)
}
\arguments{
\item{obj}{topic model object}

\item{number_words}{Top n words per topic to use.  Default is 4}

\item{separator}{The string with which the words and topic prefix will be separated. Underscores are the default but a nice alternative is ", ".  Default `_`}

\item{word_length}{The maximum length of each word in the topic label. Some words might be relatively long and setting this value helps to make sure that all labels have relatively similar lengths.}

\item{update_topic_model_labels}{if `TRUE` updates new label into a custom field}

\item{append_number_words}{Append the number of words}

\item{topic_prefix}{Whether to use the topic ID as a prefix. If set to True, the topic ID will be separated using the separator}
}
\description{
Generate BERT Topic Labels
}
