% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topic_per_class}
\alias{bert_topic_per_class}
\title{Topics Per Structured Class}
\usage{
bert_topic_per_class(
  topic_model,
  docs = NULL,
  classes = NULL,
  global_tuning = TRUE
)
}
\arguments{
\item{global_tuning}{}
}
\description{
Topics Per Structured Class
}
