% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_topics_over_time}
\alias{bert_topics_over_time}
\title{Extract Topcis Over Time}
\usage{
bert_topics_over_time(
  obj,
  docs = NULL,
  timestamps = NULL,
  nr_bins = NULL,
  datetime_format = NULL,
  evolution_tuning = TRUE,
  global_tuning = TRUE,
  return_tibble = TRUE
)
}
\arguments{
\item{return_tibble}{}
}
\description{
Extract Topcis Over Time
}
