% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_transform}
\alias{bert_transform}
\title{BERTopic Transform}
\usage{
bert_transform(obj, documents, embeddings = NULL, images = NULL)
}
\arguments{
\item{obj}{A BERTopic model object created by `bert_topic()`.}

\item{documents}{Character vector. A single document or list of documents to transform.}

\item{embeddings}{Numeric matrix or `NULL`. Pre-trained document embeddings. These can be used instead of the sentence-transformer model. Default `NULL`.}

\item{images}{List or `NULL`. A list of paths to the images to predict on or the images themselves. Default `NULL`.}
}
\value{
The transformed BERTopic model object with predicted topics for new documents.
}
\description{
After having fit a model, use transform to predict new instances.
}
