% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{bert_transform}
\alias{bert_transform}
\title{BERTopic Transform}
\usage{
bert_transform(obj, documents, embeddings = NULL)
}
\arguments{
\item{obj}{BERTopic Object}

\item{documents}{A single document or a list of documents to fit on}

\item{embeddings}{If not `NULL` Pre-trained document embeddings. These can be used instead of the sentence-transformer model.}
}
\description{
After having fit a model, use transform to predict new instances
}
