% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_transform_documents}
\alias{bert_transform_documents}
\title{Transform New Documents}
\usage{
bert_transform_documents(obj, documents = NULL, embeddings = NULL)
}
\arguments{
\item{obj}{bertopic objects}

\item{documents}{vector of documents}

\item{embeddings}{if not `NULL` matrix or dataframe of embeddings}
}
\description{
Transform New Documents
}
