% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bert_update_topics}
\alias{bert_update_topics}
\title{Updates the topic representation by recalculating c-TF-IDF with the new parameters as defined in this function.}
\usage{
bert_update_topics(
  obj,
  docs = NULL,
  topics = NULL,
  top_n_words = 10,
  n_gram_range = NULL,
  vectorizer_model = NULL,
  ctfidf_model = NULL,
  representation_model = NULL
)
}
\arguments{
\item{obj}{BERTopic Object}

\item{docs}{The documents you used when calling either fit or fit_transform}

\item{topics}{A list of topics where each topic is related to a document in docs. Use this variable to change or map the topics. NOTE: Using a custom list of topic assignments may lead to errors if topic reduction techniques are used afterwards. Make sure that manually assigning topics is the last step in the pipeline.  Default `NULL`}

\item{top_n_words}{The number of words per topic to extract. Setting this too high can negatively impact topic embeddings as topics are typically best represented by at most 10 words.  Default `10`}

\item{n_gram_range}{The n-gram range for the CountVectorizer. Default `NULL`}

\item{vectorizer_model}{Pass in your own CountVectorizer from scikit-learn.  Default `NULL`}

\item{ctfidf_model}{Pass in your own c-TF-IDF model to update the representations.  Default `NULL`}

\item{representation_model}{Pass in a model that fine-tunes the topic representations calculated through c-TF-IDF. Models from bertopic.representation are supported. Default `NULL`}
}
\description{
Updates the topic representation by recalculating c-TF-IDF with the new parameters as defined in this function.
}
