% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic-package.r
\docType{package}
\name{bertopic-package}
\alias{bertopic}
\alias{bertopic-package}
\title{bertopic: bertopic and friends}
\description{
R port of the python bertopic and other python nlp modules
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abresler/bertopic}
  \item Report bugs at \url{https://github.com/abresler/bertopic/issues}
}

}
\author{
\strong{Maintainer}: Alex Bresler \email{abresler@asbcllc.com}


Alex Bresler (abresler@asbcllc.com)
}
\keyword{internal}
