% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BERTopic-package}
\alias{BERTopic}
\alias{BERTopic-package}
\title{BERTopic: Topic Modeling with 'BERTopic' in R}
\description{
Interface to the Python package 'BERTopic' for transformer-based topic modeling. Provides R wrappers to fit BERTopic models, transform new documents, update and reduce topics, extract topic- and document-level information, and generate interactive visualizations. Python backends and dependencies are managed via the 'reticulate' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Feng-Ji-Lab/BERTopic}
  \item Report bugs at \url{https://github.com/Feng-Ji-Lab/BERTopic/issues}
}

}
\author{
\strong{Maintainer}: Biying Zhou \email{biying.zhou@psu.edu}

}
\keyword{internal}
