% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{bertopic_load}
\alias{bertopic_load}
\title{Load a BERTopic model}
\usage{
bertopic_load(path)
}
\arguments{
\item{path}{Path used in \code{\link[=bertopic_save]{bertopic_save()}} (file or directory).}
}
\value{
A "bertopic_r" object with the loaded Python model.
}
\description{
Load a BERTopic model from disk that was saved with \code{\link[=bertopic_save]{bertopic_save()}}.
}
