% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{bertopic_representations}
\alias{bertopic_representations}
\title{Import Bertopic Representations}
\usage{
bertopic_representations(obj = NULL)
}
\arguments{
\item{obj}{Bertopic Module}
}
\description{
Import Bertopic Representations
}
