% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{bertopic_representations}
\alias{bertopic_representations}
\title{Import Bertopic Representations}
\usage{
bertopic_representations(
  obj = NULL,
  numba_threads = 1,
  use_token_parallel = TRUE
)
}
\arguments{
\item{obj}{Bertopic Module}

\item{use_token_parallel}{}
}
\description{
Import Bertopic Representations
}
