% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{bertopic_session_info}
\alias{bertopic_session_info}
\title{Summarize Python/BERTopic session info}
\usage{
bertopic_session_info()
}
\value{
A data.frame with key versions and paths
}
\description{
Summarize Python/BERTopic session info
}
