% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{cohere_representation}
\alias{cohere_representation}
\title{Title}
\usage{
cohere_representation(client = NULL, model = NULL, prompt = NULL, obj = NULL)
}
\arguments{
\item{obj}{}
}
\description{
Title
}
