% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dictionary_nltk_stopwords}
\alias{dictionary_nltk_stopwords}
\title{NLTK Stopwords}
\usage{
dictionary_nltk_stopwords(language = "english")
}
\arguments{
\item{language}{language defaults to `english`}
}
\description{
NLTK Stopwords
}
\examples{
dictionary_nltk_stopwords()
}
