% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{distilbert_embedding_transformer}
\alias{distilbert_embedding_transformer}
\title{Distilbert Transformer}
\usage{
distilbert_embedding_transformer(obj = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Distilbert Transformer
}
