% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-python.R
\name{.need_py}
\alias{.need_py}
\title{Check Python availability and required modules at runtime}
\usage{
.need_py()
}
\description{
Check Python availability and required modules at runtime
}
\keyword{internal}
