% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_bert_topics}
\alias{extract_bert_topics}
\title{Extract Berttopics from output}
\usage{
extract_bert_topics(
  obj,
  docs = NULL,
  id_columns = NULL,
  text_column = NULL,
  topic_model = NULL,
  include_labels = T,
  number_words = 4L,
  arrange_topics = F
)
}
\arguments{
\item{label_words}{}
}
\description{
Extract Berttopics from output
}
