% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_bert_umap}
\alias{extract_bert_umap}
\title{Extract UMAP from Object}
\usage{
extract_bert_umap(obj, data = NULL, number_zeros = 4)
}
\arguments{
\item{data}{}
}
\description{
Extract UMAP from Object
}
