% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_document_word_counts}
\alias{extract_document_word_counts}
\title{Extract Document Word Coutns}
\usage{
extract_document_word_counts(obj, filter_zero = T)
}
\arguments{
\item{obj}{}
}
\description{
Extract Document Word Coutns
}
