% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{gather_keybert_keywords}
\alias{gather_keybert_keywords}
\title{Gather Keybert Generated Keywords}
\usage{
gather_keybert_keywords(data, id_column = NULL, only_distinct = F)
}
\arguments{
\item{only_distinct}{}
}
\description{
Gather Keybert Generated Keywords
}
