% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{gather_keybert_keywords}
\alias{gather_keybert_keywords}
\title{Gather Keybert Generated Keywords}
\usage{
gather_keybert_keywords(
  data,
  id_column = NULL,
  other_join_columns = NULL,
  only_distinct = F
)
}
\arguments{
\item{data}{A `tbl`}

\item{id_column}{Name of ID coliumn}

\item{other_join_columns}{If `TRUE` other column features to include}

\item{only_distinct}{if `TRUE` only distinct keywords}
}
\description{
Gather Keybert Generated Keywords
}
