% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{hugging_face_transformers}
\alias{hugging_face_transformers}
\title{Hugging Face Transformers}
\usage{
hugging_face_transformers(
  obj = NULL,
  task = "feature-extraction",
  model = "distilbert-base-cased",
  ...
)
}
\arguments{
\item{model}{model name options
\itemize{
\item \href{https://huggingface.co/models}{hugging face models}
}}

\item{...}{}
}
\description{
Hugging Face Transformers
}
