% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{hugging_face_transformers}
\alias{hugging_face_transformers}
\title{Hugging Face Transformers}
\usage{
hugging_face_transformers(
  obj = NULL,
  task = "feature-extraction",
  model = "distilbert-base-cased",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Hugging Face Transformers
}
