% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{import_bertopic}
\alias{import_bertopic}
\title{Import Bert Topic}
\usage{
import_bertopic(assign_to_environment = T, path = NULL)
}
\arguments{
\item{assign_to_environment}{if \code{TRUE} assigns to environment}
}
\value{
python object
}
\description{
\href{https://maartengr.github.io/BERTopic}{BERTopic API from python}
}
\examples{
library(bertopic)
import_bertopic()
}
