% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modules.R
\name{import_gensim}
\alias{import_gensim}
\title{Import Gensim Module}
\usage{
import_gensim(assign_to_environment = T, path = NULL)
}
\arguments{
\item{path}{}
}
\description{
\itemize{
\item \href{https://radimrehurek.com/gensim/}{GENSIM}
}
}
