% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{import_keybert}
\alias{import_keybert}
\title{Import Keybert Module}
\usage{
import_keybert(
  assign_to_environment = T,
  use_token_parallel = FALSE,
  path = NULL
)
}
\arguments{
\item{assign_to_environment}{if `TRUE` assigns to environment}

\item{path}{python path}
}
\description{
Import Keybert Module
}
\examples{
library(bertopic)
import_keybert()
}
