% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modules.R
\name{import_python_modules}
\alias{import_python_modules}
\title{Import Python Modules}
\usage{
import_python_modules(
  modules = NULL,
  python_path = NULL,
  assign_to_environment = FALSE
)
}
\arguments{
\item{python_path}{}
}
\description{
Import Python Modules
}
\examples{
import_python_modules(modules = c("umap", "hdbscan"))
hdbscan
umap
}
