% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{import_rapidfuzz}
\alias{import_rapidfuzz}
\title{Import RapidFuzz Module}
\usage{
import_rapidfuzz(assign_to_environment = T, path = NULL)
}
\arguments{
\item{path}{}
}
\description{
RapidFuzz module \url{https://maxbachmann.github.io/RapidFuzz/}
}
