% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{import_sentence_transformers}
\alias{import_sentence_transformers}
\title{Sentence Transformer Module}
\usage{
import_sentence_transformers(assign_to_environment = T, path = NULL)
}
\arguments{
\item{path}{}
}
\description{
Sentence Transformer Module
}
