% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{import_transformers}
\alias{import_transformers}
\title{Import Hugging Face Transformers}
\usage{
import_transformers(assign_to_environment = T, path = NULL)
}
\arguments{
\item{path}{}
}
\description{
Import Hugging Face Transformers
}
