% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{keybert_embeddings}
\alias{keybert_embeddings}
\title{Extract Embeddings from Documents}
\usage{
keybert_embeddings(
  docs = NULL,
  obj = NULL,
  model = "all-MiniLM-L6-v2",
  stopword_package_sources = NULL,
  extra_stop_words = NULL,
  exclude_stop_words = T,
  assign_to_environment = T,
  keyphrase_ngram_range = list(1L, 1L),
  use_key_phrase_vectorizer = T,
  candidates = NULL,
  language = "english",
  is_lower_case = T,
  pos_pattern = "<J.*>*<N.*>+",
  vocabulary = NULL,
  use_yake_candidates = F,
  min_df = 1L,
  max_df = 1L
)
}
\arguments{
\item{max_df}{}
}
\description{
Extract Embeddings from Documents
}
