% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{keybert_keywords}
\alias{keybert_keywords}
\title{Keybert Keywords}
\usage{
keybert_keywords(
  docs = NULL,
  obj = NULL,
  model = "all-MiniLM-L6-v2",
  exclude_stop_words = T,
  keyphrase_ngram_range = list(1L, 1L),
  use_embeddings = F,
  use_key_phrase_vectorizer = T,
  top_n_words = 5L,
  use_mmr = T,
  diversity = 0.5,
  nr_candidates = 20L,
  seed_keywords = NULL,
  doc_embeddings = NULL,
  word_embeddings = NULL,
  candidates = NULL,
  use_yake_candidates = F,
  stop_words = "english",
  is_lower_case = T,
  extra_stop_words = NULL,
  min_df = 1L,
  max_df = 1L,
  pos_pattern = "<J.*>*<N.*>+",
  use_maxsum = FALSE,
  vocabulary = NULL,
  highlight = F
)
}
\arguments{
\item{vocabulary}{}
}
\description{
Keybert Keywords
}
\examples{
 doc <- "Sources tell us that Google is acquiring Kaggle, a platform that hosts data science and machine learning competitions. Details about the transaction remain somewhat vague, but given that Google is hosting its Cloud Next conference in San Francisco this week, the official announcement could come as early as tomorrow. Reached by phone, Kaggle co-founder CEO Anthony Goldbloom declined to deny that the acquisition is happening. Google itself declined 'to comment on rumors'. Kaggle, which has about half a million data scientists on its platform, was founded by Goldbloom  and Ben Hamner in 2010. The service got an early start and even though it has a few competitors like DrivenData, TopCoder and HackerRank, it has managed to stay well ahead of them by focusing on its specific niche. The service is basically the de facto home for running data science and machine learning competitions. With Kaggle, Google is buying one of the largest and most active communities for data scientists - and with that, it will get increased mindshare in this community, too (though it already has plenty of that thanks to Tensorflow and other projects). Kaggle has a bit of a history with Google, too, but that's pretty recent. Earlier this month, Google and Kaggle teamed up to host a $100,000 machine learning competition around classifying YouTube videos. That competition had some deep integrations with the Google Cloud Platform, too. Our understanding is that Google will keep the service running - likely under its current name. While the acquisition is probably more about Kaggle's community than technology, Kaggle did build some interesting tools for hosting its competition and 'kernels', too. On Kaggle, kernels are basically the source code for analyzing data sets and developers can share this code on the platform (the company previously called them 'scripts'). Like similar competition-centric sites, Kaggle also runs a job board, too. It's unclear what Google will do with that part of the service. According to Crunchbase, Kaggle raised $12.5 million (though PitchBook says it's $12.75) since its   launch in 2010. Investors in Kaggle include Index Ventures, SV Angel, Max Levchin, Naval Ravikant, Google chief economist Hal Varian, Khosla Ventures and Yuri Milner."

keybert_keywords(docs = doc, top_n_words = 10)
keybert_keywords(docs = doc, top_n_words = 10, pos_pattern='<N.*>') # only nouns
doc_sample <-  "Supervised learning is the machine learning task of learning a function that maps an input to an output based on example input-output pairs. It infers a function from labeled training data consisting of a set of training examples. In supervised learning, each example is a pair consisting of an input object (typically a vector) and a desired output value (also called the supervisory signal). A supervised learning algorithm analyzes the training data and produces an inferred function, which can be used for mapping new examples. An optimal scenario will allow for the algorithm to correctly determine the class labels for unseen instances. This requires the learning algorithm to generalize from the training data to unseen situations in a 'reasonable' way (see inductive bias)."

keybert_keywords(docs = doc_sample, top_n_words = 10,  use_key_phrase_vectorizer = F, keyphrase_ngram_range = list(1L, 4L))
keybert_keywords(docs = doc_sample, top_n_words = 10,  use_key_phrase_vectorizer = F, keyphrase_ngram_range = list(1L, 4L), use_yake_candidates = T)
keybert_keywords(docs = doc_sample, top_n_words = 10,  use_key_phrase_vectorizer = F, keyphrase_ngram_range = list(1L, 4L), seed_keywords = 'information')
keybert_keywords(docs = doc_sample, top_n_words = 10,  use_key_phrase_vectorizer = T, keyphrase_ngram_range = list(1L, 4L), seed_keywords = 'information', use_embeddings = T)
}
