% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{keybert_keywords}
\alias{keybert_keywords}
\title{Keybert Keywords}
\usage{
keybert_keywords(
  docs = NULL,
  obj = NULL,
  iterate_individually = FALSE,
  return_message = TRUE,
  model = "all-MiniLM-L6-v2",
  stopword_package_sources = NULL,
  extra_stop_words = NULL,
  exclude_stop_words = T,
  assign_to_environment = F,
  keyphrase_ngram_range = list(1L, 1L),
  use_future = TRUE,
  use_embeddings = F,
  use_key_phrase_vectorizer = T,
  top_n_words = 5L,
  use_mmr = T,
  diversity = 0.5,
  nr_candidates = 20L,
  seed_keywords = NULL,
  doc_embeddings = NULL,
  word_embeddings = NULL,
  candidates = NULL,
  use_yake_candidates = F,
  language = "english",
  is_lower_case = T,
  min_df = 1L,
  max_df = 1L,
  pos_pattern = "<J.*>*<N.*>+",
  use_maxsum = FALSE,
  vocabulary = NULL,
  highlight = F,
  decay = NULL,
  delete_min_df = NULL,
  workers = 6L,
  spacy_pipeline = "en_core_web_sm",
  custom_pos_tagger = NULL
)
}
\arguments{
\item{docs}{The document(s) for which to extract keywords/keyphrases}

\item{obj}{Keybert Object}

\item{model}{Use a custom embedding model. The following backends are currently supported: * SentenceTransformers * 🤗 Transformers * Flair * Spacy * Gensim * USE (TF-Hub) You can also pass in a string that points to one of the following sentence-transformers models: * https://www.sbert.net/docs/pretrained_models.html}

\item{stopword_package_sources}{if not `NULL` c("snowball","stopword -iso", "smart", "nltk"))}

\item{extra_stop_words}{vector of extra stop words}

\item{exclude_stop_words}{if `TRUE` excludes stop words}

\item{assign_to_environment}{if `TRUE` assigns objec to environment}

\item{keyphrase_ngram_range}{Length, in words, of the extracted keywords/keyphrases. NOTE: This is not used if you passed a vectorizer.}

\item{use_embeddings}{if `TRUE` uses embeddings}

\item{use_key_phrase_vectorizer}{if `TRUE` uses kephrase vectorizer}

\item{top_n_words}{Number of top words to return}

\item{use_mmr}{if `TRUE` Calculate Maximal Marginal Relevance (MMR) between candidate keywords and the document. and default is `TRUE`}

\item{diversity}{How diverse the select keywords/keyphrases are. Values between 0 and 1 with 0 being not diverse at all and 1 being most diverse. Default is `0.5`}

\item{nr_candidates}{The number of candidates to consider default `20L`}

\item{seed_keywords}{Seed keywords that may guide the extraction of keywords by steering the similarities towards the seeded keywords. Default `NULL`}

\item{doc_embeddings}{The embeddings of each document. Default `NULL`}

\item{word_embeddings}{Word embeddings to use if not `NULL`}

\item{candidates}{Candidate keywords/keyphrases to use instead of extracting them from the document(s) NOTE: This is not used if you passed a vectorizer.}

\item{use_yake_candidates}{if `TRUE` uses Yake Keyword Extractor.  Default  `FALSE`}

\item{language}{Language to use.  Default `english`}

\item{is_lower_case}{if  `TRUE` lower cases}

\item{min_df}{Minimum document frequency of a word across all documents if keywords for multiple documents need to be extracted. NOTE: This is not used if you passed a vectorizer.}

\item{max_df}{Maximum number of documents.  Default `1`}

\item{pos_pattern}{KeyphraseVectorizers extracts the part-of-speech tags from the documents and then applies a regex pattern to extract keyphrases that fit within that pattern. The default pattern is <J.*>*<N.*>+ which means that it extract keyphrases that have 0 or more adjectives followed by 1 or more nouns.}

\item{use_maxsum}{if `TRUE` Calculate Max Sum Distance for extraction of keywords We take the 2 x top_n most similar words/phrases to the document. Then, we take all top_n combinations from the 2 x top_n words and extract the combination that are the least similar to each other by cosine similarity. This is O(n^2) and therefore not advised if you use a large top_n.}

\item{vocabulary}{`SKLearn` vocabulary}

\item{highlight}{Whether to print the document and highlight its keywords/keyphrases. NOTE: This does not work if multiple documents are passed.}

\item{custom_pos_tagger}{}
}
\value{
`tibble` of keywords
}
\description{
Keybert Keywords
}
\examples{
 doc <- "Sources tell us that Google is acquiring Kaggle, a platform that hosts data science and machine learning competitions. Details about the transaction remain somewhat vague, but given that Google is hosting its Cloud Next conference in San Francisco this week, the official announcement could come as early as tomorrow. Reached by phone, Kaggle co-founder CEO Anthony Goldbloom declined to deny that the acquisition is happening. Google itself declined 'to comment on rumors'. Kaggle, which has about half a million data scientists on its platform, was founded by Goldbloom  and Ben Hamner in 2010. The service got an early start and even though it has a few competitors like DrivenData, TopCoder and HackerRank, it has managed to stay well ahead of them by focusing on its specific niche. The service is basically the de facto home for running data science and machine learning competitions. With Kaggle, Google is buying one of the largest and most active communities for data scientists - and with that, it will get increased mindshare in this community, too (though it already has plenty of that thanks to Tensorflow and other projects). Kaggle has a bit of a history with Google, too, but that's pretty recent. Earlier this month, Google and Kaggle teamed up to host a $100,000 machine learning competition around classifying YouTube videos. That competition had some deep integrations with the Google Cloud Platform, too. Our understanding is that Google will keep the service running - likely under its current name. While the acquisition is probably more about Kaggle's community than technology, Kaggle did build some interesting tools for hosting its competition and 'kernels', too. On Kaggle, kernels are basically the source code for analyzing data sets and developers can share this code on the platform (the company previously called them 'scripts'). Like similar competition-centric sites, Kaggle also runs a job board, too. It's unclear what Google will do with that part of the service. According to Crunchbase, Kaggle raised $12.5 million (though PitchBook says it's $12.75) since its   launch in 2010. Investors in Kaggle include Index Ventures, SV Angel, Max Levchin, Naval Ravikant, Google chief economist Hal Varian, Khosla Ventures and Yuri Milner."

keybert_keywords(docs = doc, top_n_words = 10)
keybert_keywords(docs = doc, top_n_words = 10, return_message = T)
keybert_keywords(docs = doc, top_n_words = 10, pos_pattern='<N.*>') # only nouns
doc_sample <-  "Supervised learning is the machine learning task of learning a function that maps an input to an output based on example input-output pairs. It infers a function from labeled training data consisting of a set of training examples. In supervised learning, each example is a pair consisting of an input object (typically a vector) and a desired output value (also called the supervisory signal). A supervised learning algorithm analyzes the training data and produces an inferred function, which can be used for mapping new examples. An optimal scenario will allow for the algorithm to correctly determine the class labels for unseen instances. This requires the learning algorithm to generalize from the training data to unseen situations in a 'reasonable' way (see inductive bias)."

keybert_keywords(docs = doc_sample, top_n_words = 10,  use_key_phrase_vectorizer = F, keyphrase_ngram_range = list(1L, 4L))
keybert_keywords(docs = doc_sample, top_n_words = 10,  use_key_phrase_vectorizer = F, keyphrase_ngram_range = list(1L, 4L), use_yake_candidates = T)
keybert_keywords(docs = doc_sample, top_n_words = 10,  use_key_phrase_vectorizer = F, keyphrase_ngram_range = list(1L, 4L), seed_keywords = 'information')
keybert_keywords(docs = doc_sample, top_n_words = 10,  use_key_phrase_vectorizer = T, keyphrase_ngram_range = list(1L, 4L), seed_keywords = 'information', use_embeddings = T)
}
