% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{keybert_model}
\alias{keybert_model}
\title{Initiate Keybert model}
\usage{
keybert_model(keybert = NULL, model = "all-MiniLM-L6-v2")
}
\description{
Runs a keybert model
}
\examples{
kb <- keybert_model()
}
