% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{keybert_model}
\alias{keybert_model}
\title{Initiate Keybert model}
\usage{
keybert_model(keybert = NULL, model = "all-MiniLM-L6-v2")
}
\arguments{
\item{model}{default is `all-MiniLM-L6-v2` you can pick any model from \itemize{
\item \href{https://www.sbert.net/docs/pretrained_models.html}{sbert}
}}
}
\description{
Runs a keybert model
}
\examples{
kb <- keybert_model()
}
