% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyphrase.R
\name{keyphrase_tf_idf_to_tibble}
\alias{keyphrase_tf_idf_to_tibble}
\title{Keyphrase TFIDF to Tibble}
\usage{
keyphrase_tf_idf_to_tibble(tfidf, docs, return_wide = T)
}
\arguments{
\item{return_wide}{}
}
\description{
Keyphrase TFIDF to Tibble
}
\examples{
docs <- c('This is the first document.', 'This document is the second document.', 'And this is the third one.', 'Is this the first document?')
tfidf <- keyphrase_tf_idf()
keyphrase_tf_idf_to_tibble(tfidf, docs)
}
