% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{langchain_representation}
\alias{langchain_representation}
\title{Lang Chain Representation}
\usage{
langchain_representation(
  chain = NULL,
  prompt = NULL,
  nr_docs = 4,
  diversity = NULL,
  doc_length = NULL,
  tokenizer = NULL,
  chain_config = NULL,
  obj = NULL
)
}
\arguments{
\item{obj}{}
}
\description{
Lang Chain Representation
}
