% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{llama_cpp_representation}
\alias{llama_cpp_representation}
\title{A llama.cpp implementation to use as a representation model.}
\usage{
llama_cpp_representation(
  model = "zephyr-7b-alpha.Q4_K_M.gguf",
  prompt = NULL,
  pipeline_kwargs = NULL,
  nr_docs = 4,
  diversity = NULL,
  doc_length = NULL,
  tokenizer = NULL,
  numba_threads = 1,
  obj = NULL
)
}
\arguments{
\item{model}{Either a string pointing towards a local LLM or a  `llama_cpp.Llama` object.}

\item{obj}{}
}
\description{
A llama.cpp implementation to use as a representation model.
}
