% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{munge_bert_document_approximate_distributions}
\alias{munge_bert_document_approximate_distributions}
\title{Munge Document Topic Proability Distributions}
\usage{
munge_bert_document_approximate_distributions(
  data,
  topic_model = NULL,
  return_wide = F
)
}
\arguments{
\item{return_wide}{}
}
\description{
Munge Document Topic Proability Distributions
}
