% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{munge_bert_topics_over_time}
\alias{munge_bert_topics_over_time}
\title{Munge Topic Over Time}
\usage{
munge_bert_topics_over_time(data, topic_model = NULL)
}
\arguments{
\item{topic_model}{}
}
\description{
Munge Topic Over Time
}
