% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{munge_bert_topics_per_class}
\alias{munge_bert_topics_per_class}
\title{Munge Topics Per Class Output}
\usage{
munge_bert_topics_per_class(data, class_name = NULL, topic_model = NULL)
}
\arguments{
\item{topic_model}{}
}
\description{
Munge Topics Per Class Output
}
