% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{online_count_vectorizer}
\alias{online_count_vectorizer}
\title{An Online Vectorizer}
\usage{
online_count_vectorizer(decay = NULL, delete_min_df = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
An Online Vectorizer
}
