% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{online_count_vectorizer}
\alias{online_count_vectorizer}
\title{An Online Vectorizer}
\usage{
online_count_vectorizer(
  decay = NULL,
  delete_min_df = NULL,
  stop_words = NULL,
  numba_threads = 1,
  use_token_parallel = F,
  obj = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
An Online Vectorizer
}
