% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{open_ai_representation}
\alias{open_ai_representation}
\title{OpenAI Representation Model}
\usage{
open_ai_representation(
  model = "gpt-3.5-turbo",
  prompt = NULL,
  delay_in_seconds = 10,
  chat = TRUE,
  obj = NULL
)
}
\arguments{
\item{model}{Open AI model NAME}

\item{prompt}{Prompt.  For example: `I have a topic that contains the following documents: [DOCUMENTS]The topic is described by the following keywords: [KEYWORDS]Based on the information above, extract a short topic label in the following format:topic: <topic label>`}

\item{delay_in_seconds}{How Long to Wait?}

\item{chat}{If `TRUE` enter chat mode.}

\item{obj}{BERTopic Object}
}
\description{
See \href{https://maartengr.github.io/BERTopic/getting_started/representation/representation.html#chatgpt}{API Example} and  \href{https://maartengr.github.io/BERTopic/changelog.html#version-0141}{Change Log Example}
}
