% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{open_ai_representation}
\alias{open_ai_representation}
\title{OpenAI Representation Model}
\usage{
open_ai_representation(
  model = "text-ada-0001",
  prompt = NULL,
  generator_kwargs = NULL,
  obj = NULL
)
}
\arguments{
\item{obj}{}
}
\description{
OpenAI Representation Model
}
