% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{part_of_speech_representation}
\alias{part_of_speech_representation}
\title{Extract Topic Keywords based on their Part-of-Speech}
\usage{
part_of_speech_representation(
  model = "en_core_web_sm",
  top_n_words = 10,
  pos_patterns = NULL,
  obj = NULL
)
}
\arguments{
\item{model}{The Spacy model to use}

\item{top_n_words}{The top n words to extract.  Default `10`}

\item{pos_patterns}{Patterns for Spacy to use. See https://spacy.io/usage/rule-based-matching and https://demos.explosion.ai/matcher and https://www.kaggle.com/code/curiousprogrammer/entity-extraction-and-classification-using-spacy for more and alsohttps://spacy.io/usage/rule-based-matching}

\item{obj}{}
}
\description{
Extract Topic Keywords based on their Part-of-Speech
}
\examples{
library(bertopic)
part_of_speech_representation(pos_patterns = list(list(list('POS' = 'ADJ'), list('POS' = 'NOUN')), list(list('POS' = 'NOUN')), list(list('POS' = 'ADJ'))))

}
