% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_edit_distance}
\alias{polyfuzz_edit_distance}
\title{Calculate the Edit Distance between lists of strings using any distance/similarity based scorer}
\usage{
polyfuzz_edit_distance(
  n_jobs = 1,
  scorer = NULL,
  model_id = NULL,
  normalize = TRUE
)
}
\arguments{
\item{scorer}{The scorer function to be used to calculate the edit distance. This function should give back a float between 0 and 1, and work as follows: scorer("string_one", "string_two").  If `NULL` uses jellyfish `jaro_winkler_similarity`}

\item{model_id}{The name of the particular instance, used when comparing models}
}
\description{
Calculate the Edit Distance between lists of strings using any distance/similarity based scorer
}
