% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_model}
\alias{polyfuzz_model}
\title{Polyfuzz Model}
\usage{
polyfuzz_model(
  method = "TF-IDF",
  model_object = NULL,
  obj = NULL,
  verbose = TRUE,
  from_list = NULL,
  to_list = NULL
)
}
\arguments{
\item{method}{the method(s) used for matching. For quick selection of models select one of the following: `EditDistance`, `TF-IDF` or `Embeddings`. If you want more control over the models above, pass in a model from polyfuzz.models. For examples, see usage below.  Default `TF-iDF`}

\item{model_object}{A pretrained model object}

\item{obj}{Polyfuzz Object}

\item{verbose}{If `TRUE` prints results}

\item{from_list}{If not `NULL` vector of from list}

\item{to_list}{If not `NULL` vector of to list}
}
\description{
Polyfuzz Model
}
