% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_model}
\alias{polyfuzz_model}
\title{Polyfuzz Model}
\usage{
polyfuzz_model(obj = NULL, method = "TF-IDF", verbose = TRUE)
}
\arguments{
\item{obj}{Polyfuzz Object}

\item{method}{the method(s) used for matching. For quick selection of models select one of the following: "EditDistance", "TF-IDF" or "Embeddings". If you want more control over the models above, pass in a model from polyfuzz.models. For examples, see usage below.  Default `TF-iDF`}

\item{verbose}{}
}
\description{
Polyfuzz Model
}
