% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_rapidmatcher}
\alias{polyfuzz_rapidmatcher}
\title{Calculate the Edit Distance between lists of strings using RapidFuzz's process function}
\usage{
polyfuzz_rapidmatcher(
  n_jobs = 1,
  score_cutoff = 0.5,
  scorer = NULL,
  model_id = NULL
)
}
\arguments{
\item{n_jobs}{Nr of parallel processes, use -1 to use all cores.  Default `1L`}

\item{score_cutoff}{The minimum similarity for which to return a good match. Should be between 0 and 1.  Default .5}

\item{scorer}{The scorer function to be used to calculate the edit distance Options: `fuzz.ratio` `fuzz.partial_ratio`  `fuzz.token_sort_ratio` `fuzz.partial_token_sort_ratio` `fuzz.token_set_ratio ` `fuzz.partial_token_set_ratio` `fuzz.token_ratio` `fuzz.partial_token_ratio` `fuzz.WRation` `fuzz.QRatio` See https://maxbachmann.github.io/rapidfuzz/usage/fuzz/ for an extensive description of the scoring methods.  If NULL uses WRatio}

\item{model_id}{The name of the particular instance, used when comparing models}
}
\description{
We are using RapidFuzz instead of FuzzyWuzzy since it is much faster and does not require the more restrictive GPL license
}
