% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_save}
\alias{polyfuzz_save}
\title{Save Polyfuzz Model}
\usage{
polyfuzz_save(obj, model_path = NULL, file_name = "polyfuzz")
}
\arguments{
\item{obj}{Polyfuzz Object}

\item{model_path}{Path to save}

\item{file_name}{Name of the file.  Defaults to polyfuzz}
}
\description{
Save Polyfuzz Model
}
