% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_sentence_embeddings}
\alias{polyfuzz_sentence_embeddings}
\title{Embed words into vectors and use cosine similarity to find
the best matches between two lists of strings}
\usage{
polyfuzz_sentence_embeddings(
  embedding_model = NULL,
  min_similarity = 0.5,
  top_n = 1,
  cosine_method = "sparse",
  model_id = NULL
)
}
\arguments{
\item{embedding_model}{The Gensim model to use, this can be either a string or the model directly model to use, this can be either a string or the model directly}

\item{min_similarity}{The minimum similarity between strings, otherwise return 0 similarity}

\item{top_n}{The number of best matches you want returned}

\item{cosine_method}{The method/package for calculating the cosine similarity. Options: `sparse` `sklearn` and `knn`.  Default is `sparse`}

\item{model_id}{model id}
}
\description{
Embed words into vectors and use cosine similarity to find
the best matches between two lists of strings
}
