% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_spacy_embeddings}
\alias{polyfuzz_spacy_embeddings}
\title{Spacy Embedder}
\usage{
polyfuzz_spacy_embeddings(
  embedding_model = NULL,
  min_similarity = 0.5,
  top_n = 1,
  cosine_method = "sparse",
  model_id = NULL
)
}
\arguments{
\item{min_similarity}{The minimum similarity between strings, otherwise return 0 similarity}

\item{top_n}{The number of best matches you want returned}

\item{cosine_method}{The method/package for calculating the cosine similarity. Options: `sparse` `sklearn` and `knn`.  Default is `sparse`}

\item{model_id}{model id}
}
\description{
Spacy Embedder
}
