% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{polyfuzz_transform}
\alias{polyfuzz_transform}
\title{Transform Unseen List}
\usage{
polyfuzz_transform(obj, from_list)
}
\arguments{
\item{obj}{Polyfuzz Model Object}

\item{from_list}{Vector of Unseen Words}
}
\description{
Transform Unseen List
}
