% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-io.R
\name{select_correct_python}
\alias{select_correct_python}
\title{Select Correct Python}
\usage{
select_correct_python(path = "/opt/miniconda3/bin/python3.12")
}
\arguments{
\item{path}{Character. Path to Python executable. If `NULL`, uses default Python.}
}
\value{
Called for side effects; returns invisibly.
}
\description{
Select Correct Python
}
