% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_correct_python}
\alias{select_correct_python}
\title{Select Correct Python}
\usage{
select_correct_python(path = "/usr/local/bin/python3")
}
\arguments{
\item{path}{}
}
\description{
Select Correct Python
}
