% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{set_bert_attributes}
\alias{set_bert_attributes}
\title{Set BERTopic Model Attriubtes}
\usage{
set_bert_attributes(obj, representation_model = NULL)
}
\arguments{
\item{obj}{BERTopic Object}

\item{representation_model}{If not `NULL` representation model}
}
\description{
Set BERTopic Model Attriubtes
}
