% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{spacy_matching_pattern_dict_to_list}
\alias{spacy_matching_pattern_dict_to_list}
\title{Convert Spacy Matching Pattern to Dictionary}
\usage{
spacy_matching_pattern_dict_to_list(
  pattern = "[{'POS': 'ADJ'},\\n           {'POS': 'NOUN'}]"
)
}
\arguments{
\item{pattern}{}
}
\description{
Convert Spacy Matching Pattern to Dictionary
}
