% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-io.R
\name{stopwords_sources}
\alias{stopwords_sources}
\title{Stopword Package Sources}
\usage{
stopwords_sources(sources = c("snowball", "stopwords-iso", "smart", "nltk"))
}
\arguments{
\item{sources}{Character vector. Stopword sources to include. Options: `"snowball"`, `"stopwords-iso"`, `"smart"`, `"nltk"`.}
}
\value{
A character vector of unique stopwords combined from specified sources.
}
\description{
Stopword Package Sources
}
