% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_array}
\alias{tbl_array}
\title{Covert Arry to Tibble}
\usage{
tbl_array(data, output_type = NULL, number_zeros = 3)
}
\arguments{
\item{data}{a matrix or array}

\item{output_type}{output type for columns}

\item{number_zeros}{padding for zeros.  Default `3`}
}
\description{
Covert Arry to Tibble
}
