% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{tbl_bert_attributes}
\alias{tbl_bert_attributes}
\title{Return Tibble of BERT Attributes}
\usage{
tbl_bert_attributes(
  obj,
  return_clean = F,
  return_wide = F,
  parameter_filter = NULL
)
}
\arguments{
\item{obj}{BERTopic Object}

\item{return_clean}{if `TRUE` returns clean set of parameters}

\item{return_wide}{if `TRUE` returns it in wide form}

\item{parameter_filter}{if not `NULL` filters the attrbutes}
}
\description{
Return Tibble of BERT Attributes
}
