% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_bert_label_cosine_similarity}
\alias{tbl_bert_label_cosine_similarity}
\title{BERT Cosine Simiarity}
\usage{
tbl_bert_label_cosine_similarity(
  obj,
  exclude_outlier = FALSE,
  include_topic_number = TRUE,
  distinct_matches = FALSE,
  separate_data = T,
  return_tibble = F
)
}
\arguments{
\item{obj}{topic model object}

\item{exclude_outlier}{if `TRUE` excludes outlier topic}

\item{include_topic_number}{if `TRUE` includes topic number}

\item{return_tibble}{if `TRUE` returns tibble}
}
\description{
BERT Cosine Simiarity
}
