% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertopic.R
\name{tbl_bert_parameter_features_to_attribute}
\alias{tbl_bert_parameter_features_to_attribute}
\title{Convert BERT Parameter Table to Attributes}
\usage{
tbl_bert_parameter_features_to_attribute(
  data,
  feature_separator = "@",
  text_separator = "|"
)
}
\arguments{
\item{feature_separator}{How to concatenate features.  Defaults to `@`}

\item{text_separator}{How to concatenate text.  Defaults to `|`}
}
\description{
Convert BERT Parameter Table to Attributes
}
