% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_bert_text_features}
\alias{tbl_bert_text_features}
\title{Pull Text and Name It for bertopic}
\usage{
tbl_bert_text_features(data, id_columns = NULL, text_column = NULL)
}
\arguments{
\item{text_column}{}
}
\description{
Pull Text and Name It for bertopic
}
