% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_bert_topic_per_class}
\alias{tbl_bert_topic_per_class}
\title{Find Topics from Class via Tibble}
\usage{
tbl_bert_topic_per_class(
  data,
  topic_model,
  document_name = NULL,
  class_name = NULL,
  global_tuning = TRUE
)
}
\arguments{
\item{global_tuning}{}
}
\description{
Find Topics from Class via Tibble
}
