% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_bert_topics_over_time}
\alias{tbl_bert_topics_over_time}
\title{Extract BERT Topics Over Time from a Tibble}
\usage{
tbl_bert_topics_over_time(
  data,
  topic_model,
  document_name = NULL,
  time_feature = NULL,
  nr_bins = NULL,
  datetime_format = NULL,
  evolution_tuning = TRUE,
  global_tuning = TRUE,
  return_tibble = TRUE
)
}
\arguments{
\item{return_tibble}{}
}
\description{
Extract BERT Topics Over Time from a Tibble
}
