% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-documents.R
\name{tbl_bert_transform_documents}
\alias{tbl_bert_transform_documents}
\title{Transform Documents from Tibble}
\usage{
tbl_bert_transform_documents(
  data,
  topic_model,
  document_name = NULL,
  embeddings = TRUE
)
}
\arguments{
\item{embeddings}{}
}
\description{
Transform Documents from Tibble
}
