% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_bert_umap_label_level}
\alias{tbl_bert_umap_label_level}
\title{Extract UMAP for Visualization}
\usage{
tbl_bert_umap_label_level(
  obj,
  exclude_outlier = FALSE,
  n_components = 2L,
  metric = "cosine",
  random_state = 42L,
  min_dist = 0.1,
  learning_rate = 1L
)
}
\arguments{
\item{exclude_outlier}{}
}
\description{
Extract UMAP for Visualization
}
