% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{tbl_keybert_data}
\alias{tbl_keybert_data}
\title{Extract Keybert Keywords from Output}
\usage{
tbl_keybert_data(out, use_future = FALSE, return_message = FALSE)
}
\arguments{
\item{return_message}{}
}
\description{
Extract Keybert Keywords from Output
}
