% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybert.R
\name{tbl_keybert_keywords}
\alias{tbl_keybert_keywords}
\title{Join Keybert Keywords to data}
\usage{
tbl_keybert_keywords(
  data,
  document_column = NULL,
  obj = NULL,
  model = "all-MiniLM-L6-v2",
  exclude_stop_words = T,
  keyphrase_ngram_range = list(1L, 1L),
  use_embeddings = F,
  use_key_phrase_vectorizer = T,
  top_n_words = 5L,
  candidates = NULL,
  use_yake_candidates = F,
  stop_words = "english",
  language = "english",
  is_lower_case = T,
  extra_stop_words = NULL,
  min_df = 1L,
  max_df = 1L,
  pos_pattern = "<J.*>*<N.*>+",
  use_maxsum = FALSE,
  use_mmr = T,
  diversity = 0.5,
  vocabulary = NULL,
  nr_candidates = 20L,
  seed_keywords = NULL,
  doc_embeddings = NULL,
  word_embeddings = NULL,
  highlight = F,
  return_summary = F,
  join_to_original_data = F,
  nest_data = F
)
}
\arguments{
\item{language}{}
}
\description{
Join Keybert Keywords to data
}
