% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfuzz.R
\name{tbl_polyfuzz_matches}
\alias{tbl_polyfuzz_matches}
\title{Polyfuzz matches in a clean tibble}
\usage{
tbl_polyfuzz_matches(obj)
}
\arguments{
\item{obj}{polyfuzz objet}
}
\description{
Polyfuzz matches in a clean tibble
}
