% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top2vec.R
\name{tbl_top_2_vec_words_topics}
\alias{tbl_top_2_vec_words_topics}
\title{Textop2vecec Topic Words}
\usage{
tbl_top_2_vec_words_topics(obj)
}
\arguments{
\item{obj}{textop2vecec object}
}
\description{
Textop2vecec Topic Words
}
