% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top2vec.R
\name{top2vec_similar_words}
\alias{top2vec_similar_words}
\title{Textop2vecec Similar Words}
\usage{
top2vec_similar_words(
  keywords = NULL,
  obj = NULL,
  keywords_exclude = NULL,
  number_words = 10L,
  use_index = FALSE,
  ef = NULL
)
}
\arguments{
\item{ef}{}
}
\description{
Textop2vecec Similar Words
}
