% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{transformer_representation}
\alias{transformer_representation}
\title{Text2Text or text generation with transformers}
\usage{
transformer_representation(
  model = "gpt2",
  prompt = NULL,
  pipeline_kwargs = NULL,
  random_state = 42,
  obj = NULL
)
}
\arguments{
\item{model}{A transformers pipeline that should be initialized as "text-generation" for gpt-like models or "text2text-generation" for T5-like models. For example, pipeline('text-generation', model='gpt2'). If a string is passed, "text-generation" will be selected by default.}

\item{prompt}{The prompt to be used in the model. If no prompt is given, self.default_prompt_ is used instead. NOTE: Use "[KEYWORDS]" and "[DOCUMENTS]" in the prompt to decide where the keywords and documents need to be inserted.}

\item{pipeline_kwargs}{Kwargs that you can pass to the transformers.pipeline when it is called.}

\item{random_state}{A random state to be passed to transformers.set_seed}

\item{obj}{}
}
\description{
Text2Text or text generation with transformers
}
