% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{use_bertopic_condaenv}
\alias{use_bertopic_condaenv}
\title{Use a conda env for BERTopic (session-scoped)}
\usage{
use_bertopic_condaenv(env)
}
\arguments{
\item{env}{Conda env name}
}
\description{
Use a conda env for BERTopic (session-scoped)
}
