% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{use_bertopic_virtualenv}
\alias{use_bertopic_virtualenv}
\title{Use a virtualenv for BERTopic (session-scoped)}
\usage{
use_bertopic_virtualenv(env = getOption("BERTopic.env", "r-bertopic"))
}
\arguments{
\item{env}{Virtualenv name. Defaults to option BERTopic.env}
}
\description{
Use a virtualenv for BERTopic (session-scoped)
}
