% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_bert_viz}
\alias{write_bert_viz}
\title{Write a BERT Visualization}
\usage{
write_bert_viz(viz, base_path = NULL, viz_name = NULL, browse_url = T)
}
\arguments{
\item{browse_url}{}
}
\description{
Write a BERT Visualization
}
