% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yake.R
\name{yake_keyword_extractor}
\alias{yake_keyword_extractor}
\title{Yake Keyword Extractor}
\usage{
yake_keyword_extractor(
  docs = NULL,
  obj = NULL,
  top_features = 10,
  language = "english",
  assign_to_environment = T,
  max_ngram_size = 2,
  deduplication_thresold = 0.9,
  deduplication_algo = "seqm",
  window_size = 1
)
}
\arguments{
\item{window_size}{}
}
\description{
Yake Keyword Extractor
}
