% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_models.R
\name{zeroshot_representation}
\alias{zeroshot_representation}
\title{Zero-shot Classification on topic keywords with candidate labels}
\usage{
zeroshot_representation(
  candidate_topics = NULL,
  model = "facebook/bart-large-mnli",
  pipeline_kwargs = list(),
  min_prob = 0.8,
  obj = NULL
)
}
\arguments{
\item{candidate_topics}{A list of labels to assign to the topics if they exceed min_prob}

\item{model}{A transformers pipeline that should be initialized as "zero-shot-classification". For example, pipeline("zero-shot-classification", model="facebook/bart-large-mnli")}

\item{pipeline_kwargs}{Kwargs that you can pass to the transformers.pipeline when it is called. NOTE: Use {"multi_label": True} to extract multiple labels for each topic.}

\item{min_prob}{The minimum probability to assign a candidate label to a topic}

\item{obj}{Bertopic Object}
}
\description{
Zero-shot Classification on topic keywords with candidate labels
}
